[roll] Roll fuchsia [assembly] subpackages can have non-unique names
Updates product assembly to allow subpackages
that are not top-level package dependencies (i.e., packages only
included as subpackages) to be anonymous.

Anonymous subpackages are not included as named targets in a TUF
repo and cannot be resolved directly, via a fuchsia-pkg://repo/...
package URL, but they are resolved as part of the resolution of
another package's dependencies.

Original-Fixed: 113190
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755942
Original-Revision: 364ab39c929eda15ed0a767fc0737df0686303df
GitOrigin-RevId: 040d05dcd3ce30f0c9aba50c0e622e904cf3eac4
Change-Id: I316646d82ad1b94a11a58512ed15ca8106943638