[roll] Roll fuchsia [hypervisor] Lazily save/restore MSRs
Delay the saving and restoring of MSRs until there is a context switch.
This significantly speeds up VM exit and entry, when we do not need to
return to host user-space. This doubles the speed of the direct mode
microbenchmarks that test NoOp syscalls.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750223
Original-Revision: 8f4d67dbccdd6a65a9c20d922af49185238b81e2
GitOrigin-RevId: 9ce78a93d7dce6ecd548e45e88c98e1fae9690c4
Change-Id: I70f68d4353dafd0f4e77254a1641acd796a2f0be