[roll] Roll fuchsia [overnet] Fix behavior when host ssh closes
This fixes the behavior that when stopping the ffx daemon, the
'ssh' on the host and the sshd on target may both remain,
and remain active, exchanging keepalives, indefinitely.
(And this _will_ currently occur when using Circuit-Switched Overnet.)

When the daemon exits, its pipe will close, so the host ssh will send
an EOF control packet to the remote sshd. When the target sshd receives
the EOF packet, it will close the remote-control-runner's stdin, which
will then get an EOF. But until the Overnet socket also closes its side
(which is connected to stdout), the runner will stick around.  If for
some reason (e.g. in the current implementation of CSO) the outgoing
socket never closes, then the runner will persist, which means the sshd
will persist, which means the host's ssh will persist.

The fix is to change the runner so it exits when _either_ the stdin
loop or the stdout loop end, rather than when both end.  In that state,
no more communication will happen in any case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768588
Original-Revision: ab5ba3b81e1fe97e182d31dfafa07b427d5d2607
GitOrigin-RevId: 5c49efecee9ddbfeae06df5d867c161a26afd946
Change-Id: I574e58659a7f5a12a2ec1184115a77e5674491c8