[roll] Roll fuchsia [fvm] Unpack command verifies image length to warn of mismatch.
Because the underlying reads are preads it will silently return zeroes
when reading past the end of the image. Dealing with the resulting
images can be confusing. Adding a warning message when the image size is
not as expected.

Tested manually with a known bad and a known good image from production.

Original-Bug: b/242716290
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713682
Original-Revision: 6ec7feb19655726fde231f542a97a73f7c479938
GitOrigin-RevId: 0463c924e8800f7b4576ec1a50e99a043c69b2aa
Change-Id: Ic05f46464c53f081eb8dec0958df7a2d4663555c