[roll] Roll fuchsia [lib][fake-object] Use _zx symbols for real syscalls
This is essentially a manual revert of:
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/362573

Per discussion with Roland, using dlsym remains fragile and we should be
able to rely on the _zx* symbols now. This resolves an issue where if an
object initialized in global cxx scope made a syscall the calls to
dlopen/dlsym would hang.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768305
Original-Revision: ed51b495516b860d625350ab3121818f3a21246e
GitOrigin-RevId: fab1dc03ff18e20be8f51c46e32c708ae6bc2558
Change-Id: I5978a747c04acd88f822d446c8cc55bd179360b0