[roll] Roll fuchsia [Settings] Remove empty() for DND
The original TODO was to update to default() so it is consistent with Rust convention. But discussed offline during stand up and found the usage of this function can be replaced with a direct construction. Therefore, update the code and delete the TODO.

Original-Fixed: 67780
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766679
Original-Revision: 6cad3faf6c39068111de3790779c441353443063
GitOrigin-RevId: 5e3a66537dec821d8cc6a8555d78cba63bf25c6f
Change-Id: I16ffe88ad0dbbe7f86306fbc84a25764c8496c4e