[roll] Roll fuchsia [netemul] Log errors from devfs instead of returning
When netemul is creating a new ethernet endpoint, it watches
/dev/class/ethernet to wait for an ethertap device to come up that has
the MAC address of the device we created. In tests with multiple
ethertap endpoints, one endpoint might be in the process of shutting
down while netemul is waiting for another one to come up. This means
that while it's watching devfs, it might try and fail to interact with
that device that is shutting down.

Because it's expected that we will see potentially nonresponsive devices
mid-teardown, rather than returning an error in this situation, we
should log a warning and keep watching the directory for the device
we're looking for.

Original-Fixed: 112823, 116538, 116543, 116537
Multiply: netstack-interfaces-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770203
Original-Revision: b53265efce8cb3f9e10620cbfdfd83e788c9fb55
GitOrigin-RevId: 4edf17c98c8e55c513f70574f44af651582de40e
Change-Id: Ib39fc512ac60df18f4a9c5f981b981b586d0a151