[roll] Roll fuchsia [kms-stateless] Ignore empty filename in WatchTee
If the file watcher is activated before the optee driver is ready, it
will can return an empty filename as a result for the watcher to
process. If this isn't guarded against, WatchTee will proceed to make a
request to the optee driver, which will fail and trigger an immediate
retry.

The result is a race condition in which fshost asks kms-stateless for
the device key to unseal zxcrypt for minfs before optee is ready. After
5 retries, fshost gives up and leaves the device in a state where minfs
  remains encrypted.

Adding this check allows the watcher to continue watching for the actual
device to show up instead of trying to proceed with an empty value.

Original-Bug: b/260904369
Test: Manual
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770386
Original-Revision: d60c72907ec5a694dd4ba1e8952952caecf5a0a9
GitOrigin-RevId: b444300cd8386a91738f2aaeaf79aa8081d314f4
Change-Id: I5e52a9f8a623633f93726f4f9b27a7a173231741