[roll] Roll fuchsia [bundles] Remove non-test targets from '//src/**:tests' groups
This is a net-0 change to the set of tests included by
//bundles/buildbot/core. [1]

This changes //bundles/buildbot/core to include //src:tests_no_e2e
and //src:e2e_tests instead of the //src group.  These groups
only dep on on the various areas ':tests' groups, which should only
directly dep on tests and sub-area '<foo>:tests' groups.
Production software should not be in the direct dependencies of any
of these groups (but the tests in question _should_ dep on the
product components, to make sure that changes to production code
marks tests as needing to be re-run).

This ensures that including '//bundles/buildbot/core' does not
include subsystems that the product itself does not include.

[1] This was verified by comparing the list of GN labels for
tests in 'tests.json' before and after the change was made.

Original-Bug: 114931
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758059
Original-Revision: 84abdbe840202b3fd73cd51a78d79724115e64a5
GitOrigin-RevId: 74310889a9f5519bb246599f870a1f21256db633
Change-Id: Ifbda6b4f9c239a903ab31fcb416bddcd0c7fb02f