[roll] Roll fuchsia [dev_authenticator] Plug TestInteraction protocol in the authenticate method
The protocol TestStorageUnlockMechanism starts up in
AlwaysFailAuthentication mode but if TestInteraction::SetSuccess() is
called, it succeeds during authentication instead. This will be used in
the AccountHandler's Interaction module to talk to the Dev authenticator
and use it for testing.

Original-Bug: 109451
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764922
Original-Revision: 6d65068ec309c02bbc876e5f0b1e2812ee432ca6
GitOrigin-RevId: bf0b47065267c3e98ff3442d1bd889fa01c68837
Change-Id: Icc3bebc92349772810f1105192bead0a138b8c61