[roll] Roll fuchsia [starnix] Allow memory reclamation in binder driver
This CL replaces the previous allocator that did not correctly reclaim
memory for a new one that is inneficient (allocation are in O(n) in the
number of previous allocations) but reclaim memory.

This should be updated to a better allocator, but should help prevent
ENOMEM issues for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770862
Original-Revision: eb3f3d64654e31e84e909221f52807e22944ba4a
GitOrigin-RevId: 34e03d927fde51572c6025fd5525024e03b078c9
Change-Id: I1e45db544f0a7a737cfedf9faec55e10f89dd2f0