[roll] Roll fuchsia [ethernet] Deflake test
Because multiple ethertap devices are created and torn down concurrently
in this test, it's possible for the test to observe a PEER_CLOSED error
interacting with a device that is mid-teardown. In this scenario, keep
watching for devices rather than propagating the error.

Original-Fixed: 116578
Multiply: ethernet-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771003
Original-Revision: 178d64fac8d75f690642e18ab82d795900f511ee
GitOrigin-RevId: 0b84be8c5f10ed03a80c04cc748735dce5985611
Change-Id: I1bf28c43947cd8b67a2c5504eed0a63223f80590