[roll] Roll fuchsia [tpm] Add //src/security/lib/tpm.
This new tpm library is the Fuchsia specific interface point for
consumers of the low level TPM interface. This is an abstraction
ontop of the TCG TSS2 interface that exposes just the
functionality we need to use in the Fuchsia project through
safe Rust interfaces. The underlying TCG TSS2 interface is written
in C and we export a Rust interface via bindgen. However this
interface still is unsafe and so this high level wrapper takes
the large exported tpm2-tss-sys interface and exposes a much
smaller subset in a Rust idomatic way that prevents accidental
misuse.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759263
Original-Revision: 840268870caf5a3561500e439742366e4e1e0387
GitOrigin-RevId: e58824e712700fe1808590f2e940eb77a55161a4
Change-Id: I9b59d3d6648ea783f652c5efd2a163e81646f982