[roll] Roll fuchsia [modular] abort initialization if sessionmgr was asked to terminate.
This is an attempt to fix fxb/115009, which resulted in a use-after-free from a call to `GetAgentOutgoingServices()` within the
intitialization path of `sessionmgr`. This is the only thing I could
find after looking at the code that would result in
AgentRunner having been destroyed prior to the async initialization task
running.

Original-Bug: 115009
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770407
Original-Revision: 12202dae40ef7375fb4668712d377ebbf78127ef
GitOrigin-RevId: c929e3d179f617604c175eaa14f6de98f20f1abf
Change-Id: I567023027ae990db09370a2794b7a8794539b653