[roll] Roll fuchsia [ffx][fho] Macro hygiene improvements
There were a few places where fho macros expected things to be in the
available namespace. This makes it so they aren't anymore, using full
paths for all but one of them.

(The one unused one is `ffx_core::Injector` in the compatibility
macro, which is important because currently we auto-inject ffx_core into
all plugin deps and if this isn't used somewhere the tool will fail to
build for not using it -- this can be fixed later)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771112
Original-Revision: b149980c40238cd593c137941785a1dc712b3573
GitOrigin-RevId: 6764a7594d16a09adde399fedffe0b0704426922
Change-Id: I88de330e024d75bba4a21a905f79516caae18ab3