[roll] Roll fuchsia [fshost] fix nand device matching
When a NandDevice was processed, it caused an error in matching, because
the BootpartMatcher was tried first, which called get_block_info, which
returned an error for NandDevices.

To fix this, and to prevent similar logic bugs, break the matching
process into two steps - matching and processing. This applies some
invariants to the matching step - it's infallible, it can't mutate the
matcher, and it can't mutate the environment. Once something is matched,
we perform the fallible mutation, and cause matching to fail if we get
an error.

Test: fshost_rust_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771324
Original-Revision: 1efbe2d93d41ff001214fe309be5dd8e44c9167a
GitOrigin-RevId: b61f3fc2fc6781e1c1fbc9a4782c58d3e3d20240
Change-Id: I7554e9a9ca576a2bcbe7bb0b050a3d2a89e08241