[roll] Roll third_party/ffmpeg [third_party/ffmpeg] fix build break in "default" profile
This CL fixes a build break in the "default" profile in which
ff_toupper4 is not defined. The source code for this symbol is
included in both libcodec and libavformat. This is managed using
the "cleaning_list" in generate_gn.py, which generates the
ffmpeg_generated.gni file. The symbol is now required in both
"max" and "default" profiles, but was only included in the gni
file for "max" using the libavcodec version. Now its included
in both profiles using the libavformat version.

This started happening, because we were previously building
the "max" profiles for both x64 and arm64 (the "max" profile
was building fine), but we now build the "default" profile
for arm64.

Original-Bug: 115454

Test: build issue
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/ffmpeg/+/771051
Original-Revision: 33c1d675e49336ac11159d450cfeeb845f3942f0
GitOrigin-RevId: 492fd854e81220df0005410b620d5dab83cad372
Change-Id: Ic5852baa1d35887daea573a6964880340c074cf3