[roll] Roll fuchsia [sdk][ffx] Add schema for ffx_tool sdk element
This adds a new schema for adding separately compiled ffx subtools to
the sdk. The schema is largely derived from host_tools, but makes
explicit which file elements are intended to be the executable and the
metadata blob that defines fho compatibility between ffx and the
subtool.

Original-Fixed: 116364
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769049
Original-Revision: 846284e59a37593bfd59e09fc473ba79f887b6c6
GitOrigin-RevId: 8b5d7b91037fe5054560bdc2cf81f4f2e4e8ec8a
Change-Id: I15d5c2634c82e78302c91ce9f50d5d5e370050d9