[roll] Roll fuchsia [zither] Add support for vector syscall parameters
This change adds support for the use of `vector`s in the current FIDL
syscall model: such a parameter is decomposed as separate pointer and
length parameters.

To avoid having to generate C bindings for structs with vectors that are
not syscall request/response structs, we update the examples of this in
library zx, which Zither must be able to successfully interpret,
replacing these types with empty structs for now. Nothing currently
relies on these structs being properly interpreted at the moment, and
they are likely to be redesigned as this effort turns to the question of
how to model syscall data types in a first-class way.

Original-Bug: 110294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771363
Original-Revision: d1e21bc4cbbd49088c10ad29f7338b3859b6430e
GitOrigin-RevId: b6ea6cd9575ba934c324c5252fdc71e2566ba228
Change-Id: I46ff32d9e1b813676a430601acf80e11016cd97a