[roll] Roll fuchsia [memfs] Add client side zx::stream support
This change adds a new memfs build target that supports client side
streams. The implementation depends on zx_pager_query_vmo_stats which is
part of next_vdso and requires components that uses it to enable
use_next_vdso.

VnodeFile doesn't need to implement VmoRead because of the
supply_zero_offset in the kernel. All VnodeFile VMOs start with a size
of 0 and grow so supply_zero_offset will indicate that the zero page can
be supplied for all pages.

Original-Bug: 95299
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764249
Original-Revision: 1fddbdd0d228d08a9149a7056d986732da87c886
GitOrigin-RevId: da59490504c653be5f825c38f6eff4d914734c78
Change-Id: Ie389903dad4aed0f079cd3b79a2835a4c1605b52