[roll] Roll fuchsia [build][bazel] Use @platforms//os:fuchsia directly.
Now that both the platform build and OOT SDK rules use the 0.0.6
release of the third-party Bazel platforms repository, which includes
a proper definition for Fuchsia, use it directly.

This is a followup to the sdk-integration CL at [1] which already
did this for the SDK rules. Note that this does not change the way
our platform definitions and toolchain resolution works.

[1] https://fuchsia-review.googlesource.com/c/sdk-integration/+/758284

Original-Bug: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770888
Original-Revision: f6e25ea04fbaf58f951ef7d14b97da4c502493fe
GitOrigin-RevId: 702569bb12c7f3aa0fb2d57136d3474d809d6780
Change-Id: Ied0adbd8e197aa6bb13d351aafda7807b7930ae3