[roll] Roll fuchsia [starnix] Update binder task API related to files
The previous API was exposing the FdTable, which is not relevant for a
RemoteBinder. Instead, expose the 3 file operation that are needed:
closing a file, exporting a file, importing a file.

Also update the behaviour when accessing files on a RemoteBinder and
return an error instead of doing a panic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770890
Original-Revision: 3640fd45a5ad12ce7f3f970f99b4eab5a3093f3b
GitOrigin-RevId: 728a108fd3d4f90c7ff13d4023e4c51fc64d18a2
Change-Id: I74fd6db6fb5702b6b69a3c8244ed7c6ccd2014ae