[roll] Roll fuchsia [codec] Remove video/h264-multi decoder
Currently both the video/h264 and video/h264-multi currently construct
the same underlying decoder implementation for both amlogic and x64. The
reason for this distinction was historical and no longer required. Tests
were renamed to remove the reference to the "multi" decoder. In some
cases, the same test was the same for the video/h264 and the
video/h264-multi decoder. In those cases the test was removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769622
Original-Revision: 60d929cf2dfc978964e876c9e2d9b9abf0d398aa
GitOrigin-RevId: 34299dbd09501fd2580436a72743c4fffbf65ae8
Change-Id: I5c00191b81c5eadd6377a9d0a7ebf7d9502b8ad1