[roll] Roll fuchsia [fidl][fidlc] Create TokenPointerList
This is a standalone vector pointing to all of the token instances
inside of a given raw AST, including all "discarded" tokens that are not
represented in the raw AST proper, but only in its |tokens| side list.
The formatter has also been updated to use this new construct instead of
inspecting the |tokens| side list directly. Additionally, the |Token|
class has been updated to take an |ordinal| argument indicating its
relative position in its |SourceFile|.

These changes are necessary to allow transformation of the raw AST in a
way that the formatter can still process. The previous iteration of the
formatter was built on the assumption that it was looking at set of
tokens that were all views into the same source buffer. With the pending
introduction of transformations, this will no longer hold: new tokens
will be inserted, old ones removed, etc. The end result will be a raw
AST (and corresponding |TokenPointerList|) that will contain some tokens
from the original source, but also possibly "synthetic" tokens as well.
To enable this functionality, the way tokens are compared for ordering
and equality (a feature relied on by the formatter) must be independent
of the underlying buffers the tokens in question are sourced from -
pointer arithmetic on the underlying spans is no longer sufficient.

Additionally, a couple of other small bugs in raw AST construction and
visiting have been fixed.

Test: fx fidlc-test
Test: noop invocation of `find . -iname '*.fidl' -not -path "./out/*" -not -path"./prebuilt/*" | xargs -I {} fx format-code --files={}`
Original-Bug: 114357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770129
Original-Revision: 950ea0249ceb5721143e9d9403dd1292c258dcf3
GitOrigin-RevId: 5dc089ffe65a6c9232545f9bd5a1a4b4dcd91440
Change-Id: Ifc7be4963d78d476b1b77e3f200e79c5d1be069e