[roll] Roll fuchsia [kernel][vm] Move list node to DiscardableVmoTracker
Instead of tracking VmCowPages* in the discardable candidate lists,
track the DiscardableVmoTracker*. We will still be able to upgrade to a
RefPtr<VmCowPages> from the list node since DiscardableVmoTracker has a
raw back pointer to the owning VmCowPages.

This CL concludes the refactor to split up discardable metadata out of
VmCowPages. The size of the VmCowPagesContainer drops from 352 to 320
bytes as a result, and the DiscardableVmoTracker is 48 bytes. In other
words, the refactor saves 32 bytes of kernel metadata per
non-discardable VMO. The majority of VMOs used in practice are
non-discardable.

Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765045
Original-Revision: 4e19c4bab9d498b77b4ed3b8d67f900fc36e7d0e
GitOrigin-RevId: 62f1fc8aa10ca13741231a62ee2aac6f2139eecb
Change-Id: Ie659d2710679dd175b039469e037303b7e8d5b89