[roll] Roll fuchsia [audio] Add audio_core/v2 AudioCapturerServer
This CL implements most features of fuchsia.media.AudioCapturer. For
each capturer, we create a ConsumerNode and a GainControl for stream
gain. Ultrasound capturers will use this sample implementation, with
Args initialized to `usage = ULTRASOUND`.

The CaptureAt and StartAsyncCapture APIs are slightly tricky to
implement. The problem is that we must specify `frames_per_packet` when
creating the ConsumerNode, but frames_per_packet is decided dynamically
by CaptureAt and StartAsyncCapture. I thought of two solutions for this
problem:

1. Create a new ConsumerNode each time a new capture starts. This might
   work ok for StartAsyncCapture (at the cost of potentially higher
   latency to actually start a capture), but it won't work for
   CaptureAt, because the client can technically use a different packet
   size for each CaptureAt call (even if no clients do this in practice,
   it's allowed by the API).

2. Create a second payload buffer (call this the "source" buffer) which
   is the same size as the client's payload buffer (call this the "dest"
   buffer). We pass the "source" buffer to the ConsumerNode, along with
   a fixed packet size (10ms). The ConsumerNode will write to the
   "source" buffer. In AudioCapturer, to generate a captured packet for
   the client, we copy from the "source" buffer to the "dest" buffer. As
   we do this copy, we translate from the fixed 10ms packet size used by
   the ConsumerNode to the arbitrary packet size the client has
   dynamically selected.

This CL implements option (2), with the help of the prior
StreamSinkServer CL, which implements the logic to copy from the
"source" buffer to the "dest" buffer via
StreamSinkServer::CapturePacket.

Original-Bug: 98652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769193
Original-Revision: 2c61bec3bad0e3dd9140c513742cf9823b037146
GitOrigin-RevId: e29ff505b07782281b34af10b40609828224b13b
Change-Id: I75ea7b88f666b79d6f0ac150e6de35bc317137ad