[roll] Roll fuchsia [fshost] support empty zxcrypt in integration tests
In the integration test fixture, support formatting the data partition
as zxcrypt with no filesystem inside. This changes adding zxcrypt to be
more additive - instead of an option to remove zxcrypt, we add a builder
method to add it in. Because most tests want to make that decision when
adding a data filesystem format, add an argument to that builder method
for adding zxcrypt.

Also add a fixture method for testing for the test file on the data
partition, to clean up a common flow in tests.

Original-Bug: 87511
Test: fx test fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771325
Original-Revision: d19904afedd2d0eb1c27636c213a96391e79f6e5
GitOrigin-RevId: 43c12ffb467a77f93763ec0a5917b939750c0622
Change-Id: Idb8f4450d5f982aa08b96c0d1c2d47d954be43de