[roll] Roll fuchsia [microbenchmarks] Generalise plumbing for same/different-CPU variants of tests
In an earlier change, plumbing was added so that the "ChannelPort" IPC
round-trip test could be run with same-CPU and different-CPU variants
(i.e. with threads pinned to the same or different CPUs).

Generalise the way that the parameters are passed so that this can
work for any of the test cases that use the ThreadOrProcess helper
class.

However, we don't enable that for the other test cases for now, to
avoid creating excessive numbers of test variants.  This plumbing just
allows interested people to see the results for other test variants by
adding additional calls to RegisterTestMultiProcSameDiffCpu().

Original-Bug: 45802
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765048
Original-Revision: 8c8dee9ee03bff2312d651c34a38efc80becddbb
GitOrigin-RevId: 638adb3aec80231ae3c3d4e09be3822487a1a86c
Change-Id: I3d2e0f3378494e9a732355633dd9a2662a5ad2cd