[roll] Roll fuchsia [fidl][cpp] Ensure server endpoint is closed before CloseHandler
Since the CloseHandler name implies that the connection is closed, it
makes more sense that we ensure that the server endpoint is gone.
Previously the server endpoint would be closed right after CloseHandler
returns. This CL makes that change and adds a test.

It's unlikely that users are relying on the specific ordering.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771536
Original-Revision: 9797273af3e79c7fc40dd9837b280d64f2aab06f
GitOrigin-RevId: 4edc9dd306ae271d32cfef6669dc5888063e3643
Change-Id: Iafeb6772aed2076c39ff1dd4bc46c860fde37a43