[roll] Roll fuchsia [memory] Add the list of VMOs per category.
This is useful to understant how VMO are grouped.

Also change the test to make sure no entry is missing
in `processes`. Provious code would pass with an empty
list.

With this change the --debug_json becomes only useful
for debug as all the information appears in the machine
output.

Original-Bug: 116229
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/772682
Original-Revision: bcbfcfe5200bbea3bf0580f367bf5913a728f930
GitOrigin-RevId: ca128671d42d9990f4b959f3f845b78213128865
Change-Id: I57a2b77ffcc0f20294c4653f4d69625068effa2c