[roll] Roll fuchsia [emu] Add a --stage flag to `ffx emu`
This CL adds a flag which delineates the `ffx emu start` subcommand at
the end of the Staging phase. Like --dry-run for Configuration, this
flag instructs `ffx emu` to short-circuit the start subcommand and print
the intermediate results of the staging. All file operations are
completed as part of the staging step, such as embedding SSH keys in the
ZBI and resizing the FVM as they are copied into the emulation instance
directory.

Original-Bug: 103121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764826
Original-Revision: 70b140844721c24ced90e7e84e4d234fbf6a943f
GitOrigin-RevId: a52fdbd8e9f5bf22606443dc22fd6198d38696ed
Change-Id: I0a173c83f4787edd956a10e2f4f09af871da7d08