[roll] Roll fuchsia [emu] Ensure --commit bypasses PBM processing
This CL adjusts the code in pbm::make_configs() to explicitly bypass the
product bundle processing during configuration when a file is provided
by the user with the --config flag. Processing a product bundle in this
scenario is wasteful and inaccurate, since none of the data in that
product bundle are going to be used.

To address the discrepancy between the passed-in configuration and the
contents of the EmulatorConfiguration, we also add a value to the
EmulatorConfiguration to indicate when the --config flag was used. When
calling `ffx emu show`, warnings are printed to explain the potential
discrepancy to the user, and some outputs are refused altogether.

Original-Bug: 100653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764827
Original-Revision: c4b878cb48404a30e6f408e4aff228aa0550c847
GitOrigin-RevId: 2a4483bdc4d9f7cf6c45b4a2ff9f175108b7b139
Change-Id: I9b42bbaa5c244f95d275d92216f766d1d07028c4