[roll] Roll fuchsia [emu] Fix --reuse to reload previous config
Previously, `ffx emu start --reuse` would only reuse the disk images
from the previous instance of the emulator. In this CL, we modify the
start-up code to read the previous instances serialized engine from disk
and reuse the configuration as well. This means the user does not need
to reapply all of the command line arguments, and the launcher does not
need to re-process the product bundle metadata anymore.

Original-Bug: 100653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764828
Original-Revision: c487f0459babaee847035e671bef55bcc490bd6f
GitOrigin-RevId: 750cab12829e7bec0deec7081f538e5f75aad27e
Change-Id: I5fef4f987aab6acfe1ea36134efc2644688a97ac