[roll] Roll fuchsia [emu] Update emulator state on deserialize
This CL ensures the most recent status is displayed for an emulator
instance. Specifically, if an emulator was most recently in the
"Running" state, but the process has since terminated, this will update
the serialized state file to indicate the emulator is now in the
"Staged" state. This reuses the existing "is_running()" code to verify
the process is still active, and only updates and writes the serialized
file on termination.

Original-Bug: 112577
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768142
Original-Revision: 322b349ce70cd476fe505f6a4ca1d76583995085
GitOrigin-RevId: 2a61a551f841c0df97ac4c1169834d96a6b8193e
Change-Id: I743b1cb1d5d0669c3b5afebcf09ebfd117b20c52