[roll] Roll fuchsia [libc] Test globals for dynamically loaded libs are registered correctly
The dynamic loader is required to call __hwasan_library_loaded from the
hwasan runtime whenever a dso is loaded in order to register global tags
into shadow memory. This is done explicitly in the loader (via
__sanitizer_library_loaded) rather than
via a function call that accepts a list of globals that is called in
.init_array to resolve an order of initialization problem. With
constructors we can encounter the following problematic scenario:

 1) library A depends on library B and also interposes one of B's symbols
 2) B's constructors are called before A's (as required for correctness)
 3) during construction, B accesses one of its "own" globals (actually
    interposed by A) and triggers a HWASAN failure due to the initialization
    for A not having happened yet

This adds tests for asserting various situations in which globals can be
accessed via module constructors work correctly without crashing with
hwasan.

Original-Bug: 57112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724143
Original-Revision: f4f76f22d6517653bf16507707cb77db616756e7
GitOrigin-RevId: b20230a5b776f6d56185a8ebc3675316f806c5cf
Change-Id: Ic83ab1c76a3267b7159d93a261f63435b2bec2bb