[roll] Roll fuchsia [zither][zx] Add support for @wrapped_return
@wrapped_return annotates a singleton, response body of a protocol
method used to define a syscall and indicates that the return type of
the syscall is actually the type of the wrapped parameter. The reason
for this workaround is that protocol methods must return a struct.

This sidesteps the need for any of heuristics employed by kazoo today.

Original-Bug: 110294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771364
Original-Revision: 825bac9ef800cb3cd301111597a8605df3b4c1ef
GitOrigin-RevId: d8c12d23711676a431d0c934dfcd874bebec0c16
Change-Id: I88d5ae21cb76b47d4061106e7f978cc20f494391