[roll] Roll fuchsia [ffxutil] Wait a little after ffx stop to allow shutdown to complete.
Sometimes the ffx.log is still being written to even after the daemon is
stopped. Since we upload this with the CAS outputs on failures, this can
cause the bot to die because the file is being written to while it's
being uploaded. See
https://chromium-swarm.appspot.com/task?id=5f01becebb5f8310. We used to
try to pgrep for the ffx daemon to check if it's still running, but
pgrep is not available on the host anymore.

Original-Bug: 92296
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771904
Original-Revision: f552e629fa4bb5f26868f267f55fa714a6e82db8
GitOrigin-RevId: e541f212b6b73cb641d20e06200b163e3a50c4e2
Change-Id: I841ce159a393204882c85e396fcbb69b00d704a0