[roll] Roll fuchsia [settings] Fix generation of migration file on first boot
On a fresh boot, when no data exists for settings, the update to write
the settings migration file is skipped over, and the initialization
logic did not write the migration file. This caused the migration file
to never be written and following builds to also be stuck in that state,
because the new fidl storage code would never find the old data, always
starting from scratch.

With this change, the migration file is written when no migrations are
run and the default values are used. This update also removes the panic
on unknown migration id, which can cause a boot-loop if an update fails
and the device falls back to the previous build.

Finally, this bug prevents migration errors from causing the service to
exit early. Instead an error is logged and the service can continue to
run.

Original-Bug: 116902
Test: Tested migration on fresh boot and on device with modified
settings. Red-green testing ensured fixes work as intended.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769862
Original-Revision: 57c577f5762048285f5eda621445a986600b4bb1
GitOrigin-RevId: 61ed16dd8385fdddea47d0c5a414bc43f29f6e3a
Change-Id: I16c540a1d5925e0e3b6f2c07a3808004cb956f00