[roll] Roll fuchsia [zither][kernel] Add (syscalls|kernel).inc entry support
This change introduces support for the macro-friendly "C decl"s of the
syscalls that we find in the kazoo-generated syscalls.inc and kernel.inc
(as well as all of the variations on cdecls.inc).

With this change, Zither's internal conception/encoding of the syscall
interface should be complete.

Original-Bug: 110294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771366
Original-Revision: 14c045c7c86760fa9b2463a01130e69345f5f677
GitOrigin-RevId: cb9b926656a32dee7a4e1c62d1030c2d026ef732
Change-Id: Id9b572e405d8e329c53bf1fbe34eed14fe188146