[roll] Roll fuchsia [testing] Add "CRASH" summary.json status
The mobly test "ERROR" status can map to the resultdb "CRASH" status to
indicate that the test crashed in some way that may not be the test's
fault. This may also be useful in the future for tests that fail due to
infrastructure reasons, e.g. the target becomes unresponsive.

The primary consumers of summary.json (recipes and Sponge processor) are
already resilient to new test statuses being added; any test with an
unknown status will be considered to have failed, which is fine in this
case.

Original-Fixed: 116847
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773164
Original-Revision: dc077b11a2d4734de23b28f7249eeee0c93cfa30
GitOrigin-RevId: c3f1676e221bc823b5468dd0acf654526b49e725
Change-Id: I1b510a18d7f4afda0f83ee3af3741a003e5c658c