[roll] Roll infra/recipes [btu] Set RELEASE_VERSION env var on release builds
When a build is triggered on a release tag, the tag will now be
converted to a release version and passed as an environment variable to
the build script so it can be used by the build process.

This is just a short-term hack to get SI releases working; once SI is
using Bazel, we can update `bazel_build_test_upload.py` to pass through
the release version via `bazel build --config ...`.

Original-Bug: 116482
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/772904
Original-Revision: 07a30d317c4bba36429235b294f6baef409a2f78
GitOrigin-RevId: 747b02ea3bca4e5001be0aa2dd6b700f7137aad1
Change-Id: Ie4888323a3448bca427ce151b3ebe4f383c359ec