[roll] Roll infra/recipes [nsjail] Forward signals to subprocesses
By default nsjail sends a SIGKILL to the subprocess as soon as it
receives any termination signal, which doesn't give the subprocess time
to gracefully clean up. This is problematic for fint, which emits output
metadata during a cleanup step in case of failure; now that it's wrapped
by nsjail, a SIGTERM sent by the recipe when a fint step times out gets
converted to a SIGKILL by nsjail, preventing fint from writing its
output metadata.

With this flag, available in nsjail 3.1 and later, nsjail will forward
the SIGTERM instead of converting it to a SIGKILL, giving fint time to
write its output metadata in case of a timeout.

Original-Fixed: 111514
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/737443
Original-Revision: f901aaeda9d0b83befc7a451a67408129f4ee3ab
GitOrigin-RevId: f78d7a9696bdbc639901362df918eeb69c5be520
Change-Id: I2a53b600e3e12a4835f79df6dbf13cc510f14d6f