[roll] Roll fuchsia [emu] Fix scoping issue between stages in emu
The recent changes to introduce a firm state machine in `ffx emu`
included additional output of interim command lines for debugging. WE
missed a scoping issue, where one of the interim values was printed and
then discarded instead of continuing on to the next stage, which was
resulting in some of the configuration getting lost before final launch.

This CL fixes the scoping issue, such that the staged files and port
maps are used by the final launch.

Original-Bug: 116956
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773546
Original-Revision: b658298bdf3583c9068448792af332110eef79b4
GitOrigin-RevId: e794fef55b24ca1c8dd66215dc03d85a047a73f5
Change-Id: I3624cf42bfcb1eed495f35ec8ca061e20710bf18