[roll] Roll fuchsia [run-boot-test] Fix matching bug and clarify build requirements
This change updates `fx run-boot-test` to state its build requirements
more helpfully and also to update the fuzzy matching logic to be more
helpful. In particular, we fix a previous bug that would view an exact
name as ambiguous if it appeared as a substring in another name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768770
Original-Revision: 401ff1edabbecbc82e3b6c2e7aab925d4c35bdb9
GitOrigin-RevId: 2259314b47d01a4e7f4f3c94037fb3bfbd6827ae
Change-Id: I1876cc83b5d3503550cf848a3f2eccc3e6ad0f37