[roll] Roll fuchsia [fidl][fix] Change "fix" to "fixable"
We'll use the term "fixable" to name the description of something we can
automatically change, and "fix" to name the actual class that does the
transformation specified by a given fixable.

I've also expanded the information associated with a fixable: it now
tracks required experimental flags, as well as the "scope", or depth of
information needed to perform the fix.

Test: fx test fidlc-test
Original-Bug: 114357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/772222
Original-Revision: 4d3c49564193b2f36f29e4ba4148fadaa14049de
GitOrigin-RevId: e8e956dc9d68207d6dc2844afc5a9311b462045e
Change-Id: I8e3e8e2c9a1c8eb787a9b5a6bcd4e171e6061c0d