[roll] Roll fuchsia [assembly] Infer blobfs compression from package mode
The blobfs is not compressed if we place the FVM in the ZBI, because the
ZBI compression will optimize the size instead. Instead of requiring
boards to provide this variable we will infer its value from the package
mode passed to image assembly.

Original-Bug: 117111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773142
Original-Revision: c0e743a3da7070dd2d694d1d986bdf5038f608fb
GitOrigin-RevId: 39895a20ffe5e212a5ff67c39d06f57d9b001bfb
Change-Id: Iecb7612de2ea09e9ea7d1216112d6797ebd20372