[roll] Roll fuchsia [sampler] Restructure sampler config to use lazy metrics.
Originally the sampler-config library would directly attach non-lazy
properties to the top-level component Inspect node, keep some references
to those properties, and mutate them over time.

With a large number of selectors, this causes Sampler to fill up its
entire Inspect space. Not only does this cause us to lose metrics added
later, it also commits the whole 256K of the Inspect hierarchy in
memory.

This change refactors some areas of Sampler to instead keep references
to the original data and lazily instantiate a larger Inspect VMO upon
request:
- Some values in the incoming config are changed to Arc so they may be
  shared.
- The counters for metric uploads are changed from Inspect properties to
  atomic counters.
- A new method on SamplerConfig attaches a lazy node to a node
  explicitly, rather than implicitly attaching to the root of the
  component Inspect hierarchy. The implementation of this lazy node
  iterates over all of the atomic counters.

Generally, the sampler-config library creates a data set that both
matches the incoming format (using serde for deserialization) and is
intended to be mutated over time (such as with SelectorLists). This
resulted in one clone over the selector lists so that projects
themselves have the ability to mutate what they need, but there is room
for improvement overall.

This is a pre-factor for bug 116888.

Original-Bug: 116888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774012
Original-Revision: a414b95d677ae94134b1757709481ba5f28dfa43
GitOrigin-RevId: df3dcd4ee068b4d0f3cf044cc98b0a506153bf28
Change-Id: I395bcf383ddb7714a16c26ca96e86cd197cf560a