[roll] Roll fuchsia [ffx_trace] Add category group support to ffx trace
The goal of category groups is to make it easier to work with
large or unfamiliar lists of trace categories. It allows maintainers
to add predefined curated tracing groups that can be used by passing
the name prefixed with # in the category list provided to ffx trace.

The implementation uses ffx config to store the information about
category groups. New groups can be added to the global config by editing
data/config.json in the trace plugin.

Original-Bug: 115758

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761184
Original-Revision: 47971c5086695bf912fb82f4f2ef089954dcf850
GitOrigin-RevId: eb1598d0d7c91bfd61506b8b7142a8060e5c16e4
Change-Id: I749535381d1f4ba57c65374228e47804bf4d7a41