[roll] Roll fuchsia [graphics][compute][spinel] Support .deviceID of 0
The Lavapipe driver is unique and has a .deviceID of 0 so update the
Spinel benchmark logic to default to the first physical device when
both the .vendorID and .deviceID are both 0.

Also a small edit to to core.h.

Tested benchmark locally with the Lavapipe driver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773165
Original-Revision: 2c1f753e3aed789d3c51ce7dece1273c37293d1f
GitOrigin-RevId: c2f0b6f242571bdb16387d554f8c808316e6a381
Change-Id: I856d8878b58774a3d45221521ca2f806f6635f44