[roll] Roll fuchsia [as370] Delete AS370 board file and boot shim
There is now a board file for Visalia, so delete AS370. This was
originally for the AS370 dev board, which is no longer used. The
package-image script for Visalia is copied to a new location and will be
picked up by a subsequent change to that board file.

Original-Bug: 94099
Test: Flashed and paved Visalia, device-enumeration-test passed
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709024
Original-Revision: 8b368ea5b051a0bafd5337c8ed08f6d28000b618
GitOrigin-RevId: 1ce9c74539527641b791eba805469f94cbbdd7d7
Change-Id: I78f80eb5af233d26e487302f1dabb2048978d2e0