[roll] Roll fuchsia [camera] Avoid calling GetDeviceInfo twice
GetDeviceInfo to camera controller was being called twice using two
separate connections. Creating a 2nd connection means that the 2nd
connection is vulnerable to the 1st connection not necessarily
completely disconnected. Also, there is no reason to call the same
FIDL call twice when expecting the results to be the same. The call
is a query for static information. One call is enough.

Original-Bug: fxb/111252
Test: Run 500 cycles to ensure zero SetUI camera probe failures.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773944
Original-Revision: d0af6714243ab17911167292015215004687f985
GitOrigin-RevId: 3102a7be605e7b93afbcad5684e5ce62567ffe80
Change-Id: I7318447e0cd335a1a5a16a7a7e9573097b2bc3f9