[roll] Roll fuchsia [assembly] Enable account partition using separate assembly input
The GN flag `include_account_in_fvm` will be used to generate a separate
account config to be passed to assembly. When this config is not
supplied, the account will not be added.

This change allows us to declare explicit GN targets for all the images
config, and point to them in the board definitions.

Original-Bug: 116083
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773723
Original-Revision: cb80c8665701e149cecdff58e92c8440863a16ba
GitOrigin-RevId: 837b7c5bdadd4191a937658e6657459cada20ed9
Change-Id: If8979a7b8f46686491b078699d37aaeb6ab840c3