[roll] Roll fuchsia [trace] Acquire a Context when Checking if Enabled
When we enabled tracing, we
1) Enabled the tracing state
2) Set up the context
3) Enabled getting a context

This means if a trace provider quickly started tracing after observing
g_state being started, it could attempt to grab the trace an event,
which would try to grab the context, find that it's empty, and then bail
silently.

This changes trace_is_enabled to check that it's able to get a context
before indicating that the trace has started. So that callers don't try
to quickly emit a trace event before the system is actually ready.

Original-Fixed: 114682
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762169
Original-Revision: 1d0104904c9d9cbb548736720505994988809b61
GitOrigin-RevId: f5197dc5dec972d0950c3fcdb57cfccd5503705a
Change-Id: Iec61f8b69787ba2fe6b6a32904c8fc76ef542d90