[roll] Roll fuchsia [kernel][x86] refresh the intel microarchiture list with newer entries
Just an addition of a few more newer (and a couple we had missed before)
to the list of intel microarchitectures. Nothing really new, except
Raptor Lake and Tremont, which is pointed at the microarch just before
it. Saltwell is added for completion, but it's unlikely the kernel will
run properly on it since it's a x86-64-v1 core.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762296
Original-Revision: 366070dfe24aa4be696661b94585afc1d85d099d
GitOrigin-RevId: d8a71d6678e2e23d6cae2990e372e4101534406b
Change-Id: I53d52aa72f528c93425a8bb6485fed076486d36f