[roll] Roll fuchsia [fidl][docs] Update Wire* type docs
Update the docs for the new C++ bindings to indicate that
fidl::WireRequest and fidl::WireEvent are type aliases, and refer to
fidl::WireResult instead of fidl::WireResponse (which is mostly not
user-visible). Provide a clearer description of what is available in
fidl::WireResult.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773625
Original-Revision: 34c0196fd0932dc703f1e33f19c9be1238309357
GitOrigin-RevId: 374408be5efe27e4a3048c23414067626649b2ec
Change-Id: Ia3cf6da8784fecd8c9e9129198edbdf1439dfd52