[roll] Roll fuchsia [base_resolver] Support anonymous subpackages
Anonymous base subpackages are added to the on device base packages
manifest (the data/static_packages file of the system_image package)
with name/variant "_do_not_use_${HASH}/0". From the perspective of the
device, these anonymous packages are now base packages and so will
appear in pkgfs/packages and pkgfs/versions. This allows the
base_resolver to resolve anonymous subpackages because now their hashes
will appear in the BasePackageIndex when it does the reverse lookup and
their package directories can be opened from pkgfs/packages. As base
packages, they will be protected from GC.

The pkgfs/packages/<dir> subdirectory of packages is now a reserved
package name. None of the package resolvers will resolve a package with
this name. (The name itself is subject to change, but is currently
chosen to clearly convey it should not be used, since this directlry is
likely to go away very soon.)

Original-Bug: fxbug.dev/101492
Original-Fixed: fxbug.dev/101213

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769753
Original-Revision: 73bc6452ece7995265a0d2746ca172d1c50a6483
GitOrigin-RevId: 2c1c9f193777b35be462701d0e0da12db71383a8
Change-Id: I2949ec0c594fa2417660f3c9b35e54c760a5668c