[roll] Roll fuchsia [fs_test] Remove mmap append-only test
Mmap'ing a file that was opened in append mode is allowed by POSIX.

ext2, ext3, and ext4 support an append-only attribute which only allows
the file to opened in append mode when writing. Files with this
attribute can not be mmap'ed with PROT_WRITE. Fuchsia does not support
this attribute and this attribute was not being used in the test.

Original-Bug: 63987
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773666
Original-Revision: cb74dfa08fa2def3224fa9dae2b066f5529b93ca
GitOrigin-RevId: 7a0e73bf3c3d1225059ce893a7abe99b7ce2f082
Change-Id: I5af7296ef0886dee7f7cf4aa978139860817edf7