[roll] Roll fuchsia [archivist] Fix bug in event stream v2 handling which would lead to unhandled events
During shutdown, we wouldn't drain the merged stream that was created in
EventSource. By creating one EventSource per event stream we ingest, we
take advantage of the logic in place in EventRouter that ensures that we
drain all the events already received when shutdown is requested.

Even with that fix (in PS3), I think there's still potential for flaking
due to a race while pushing events through `dispatcher.emit().await` and
the receiver being terminated. To remove this race, I made the channel
unbounded so that the code that  places the events into the internal buffer is
entirely syncrhonous removing the potential of the receiver going away
while we are placing events in the channel. I don't think making this an
unbounded channel will be a problem in practice since it'll be storing
handles and identities that we'll be storing when the event is processed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775103
Original-Revision: 990127041deee83c14eaebbe8825dbe83a0d30a6
GitOrigin-RevId: 7362b2f217a4d6a5fe409b03b803c12b4985746d
Change-Id: I99736661605bca24ee93ca8e064e1158877dab7f