[roll] Roll fuchsia [build][sdk] Add missing variants to ElementTypes
This adds the following variants to the ElementTypes enum that are used
in the sdk but weren't present before. This was causing failures in ffx
trying to read the released manifest after switching to using this
library to parse things.

It also adds a catch-all Unknown variant that will avoid erroring on
future additions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775243
Original-Revision: ab7cef8f1769bfbb0608617e7bf69b8a55ae35b6
GitOrigin-RevId: ed6f384d22ba25417b7853817a677dfd98cbbff6
Change-Id: I7369bc80a5dd9d16fad24245510ad7880a4106d6