[roll] Roll fuchsia [ffx][fho] Add writer infrastructure
This adds a Writer associated type to FfxMain and uses it as an
additional argument to the main() function. Originally we intended this
to be something that would be included in the FfxTool struct, but with
the goal of minimizing complicated macro machinery it seemed better to
move it out to its own thing.

A new trait is introduced, `FfxToolIo`, that currently only mimics the
`ffx_plugin` functions around machine output, and defaults them to
saying no machine output. A followup change will add a concrete type to
the ffx_writer crate that requires a concrete machine type and will impl
this trait with 'better' definitions for those things.

Original-Bug: 114369
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774662
Original-Revision: 8f2d284abffcec7bf0fdae412378b184a73d5184
GitOrigin-RevId: 30dac80f3adedd5e5d7c80eb380bec6fd82a5282
Change-Id: Ied316cfbab72300ad9b4685b801d61437f8b42a7