[roll] Roll fuchsia [starnix][ext4] Use exclamation mark as separator for symbolization
The VMO name for ELF segments is used in the symbolizer output to
identify modules. The symbolizer format does not support using ':' or
'}' in the name field and does not have an escape mechanism:

https://fuchsia.dev/fuchsia-src/reference/kernel/symbolizer_markup

This uses 'ext!' instead of 'ext4:' as a prefix to work around this
limitation. This lets tools that parse symbolizer output process these
modules correctly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775622
Original-Revision: 3025b3e201a1083719adf7f70d856597c5c60b3b
GitOrigin-RevId: 411cab15b64011f320f06687ecf51058d455e5e3
Change-Id: Icbc2f076edd05345aaa678594ce4e57716e78f44