[roll] Roll fuchsia [debugger] Cleanup dead code, fixes.
Removes the lazy symbolization in the FrameImpl class. This was leftover
from when the Stack didn't symbolize the locations. This code path
hasn't been used in some time.

Update the MockModuleSymbols to mark failed matches as still Symbolized
since it has gone through the symbol lookup process. This contributed
triggering the below-described test problem.

Update the debug adapter request_backtrace test to account for the
function call sites changed for issue 96999. This change meant that the
previous address was looked up rather than the exact one given in the
stack. It turns out the exact codepath used by this test triggered the
lazy symbolization codepath (now removed) which didn't have the fix for
the function call sites in it, masking the bug. The test is updated.

Remove the explicitly defaulted Location destructor which prevented the
default move constructor from being generated. A new #include was needed
in the symbol system to use this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775523
Original-Revision: 75421ff8934f51281d58e9e15c1a0e7204d40606
GitOrigin-RevId: bf7dc2443ce7d81a5813386045a500bf6620b25d
Change-Id: I4c8a30234e218fd604dfb825f80b8ba548c8822f