[roll] Roll infra/recipes [nsjail] Add git-upload-pack to the sandbox
`git-upload-pack` is a separate binary from `git` despite being a
built-in command, so it must be added to the sandbox separately.

It's invoked in Bazel actions that Bazel SDK workflows, some of which
clone a repo from a local remote, and the cloning processing calls `git
upload-pack` even when the remote is local.

Original-Bug: 117237
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/775264
Original-Revision: 50175fee26e8d00850e44af743d1375e530d3d5f
GitOrigin-RevId: 7a75a0c1fe244351cb0a5006e2bda443dc7c9679
Change-Id: I6db4d4678391914d3b2e67e463e9ba00659d04b2