[roll] Roll chromium/deps/icu@default [fuchsia] Limit the visibility of ICU targets
This narrows down the code that uses ICU in Fuchsia to a very specific
subdirectory of the build tree.

Fuchsia needs this because it requires very specific choices
in the ICU library, and we use visibility to force all
dependencies to take in the code that makes those
choices correctly.

Original-Bug: fuchsia:115614
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4035094
Original-Revision: 9a2f72aa54e09196b39727972b9efae2dcb6a170
GitOrigin-RevId: c52d5d16a191fb6076a8000cc54503ab527dec7f
Change-Id: Ia93bedcccdb236f2e6b3dfc197756c32f879a431