[roll] Roll gwp-asan [GWP-ASan] Fix sanitizer backtrace util using wrong print.
The sanitizer backtrace is used in tests, and calling StackTrace.Print()
means that it uses the sanitizer's Printf(), rather than GWP-ASan's
Printf(). In the current code, GWP-ASan's Printf() *is* the sanitizer
print, but this isn't guaranteed to be the case, and will change in an
upcoming patch.

Reviewed By: eugenis

Differential Revision: https://reviews.llvm.org/D139748

GitOrigin-RevId: 8bde7ad34e2568acdef1b286a8fed141339a5a6d
Original-Revision: 2ad1a5e48efec766678f661b5206e35fa9ab5008
Change-Id: I2f2102d5face7729edfab5d3f3a7989825de1662