[roll] Roll fuchsia [flatland] Use prunable tokens in DisplayCompositor
Replaces AttachToken logic with prunable tokens. This lets us
include the display in sysmem negotiations, but with the option
for it to be excluded if there is no way to complete negotiations
with it involved.

Original-Fixed: 111107

Multiply: flatland_engine_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759269
Original-Revision: 98a04a26f4750c47c52dcb22e566af2b38bc73dc
GitOrigin-RevId: 8a00e3667be9c7c27c0ee63e37d9202c89aa4307
Change-Id: I0ff7cdba9322f887a63779e6916525d3286e7fc0