[roll] Roll fuchsia [kernel][vm] Batch page requests to page sources
When performing direct operations on a VMO, the kernel knows exactly the
range in which pages are needed. For pages with page sources, this
change enables sending page requests in batches instead of separately
page-by-page. Consequently, when performing an operation on N
non-present pages of a pager-backed VMO, this reduces the number of the
outbound page requests to the userpager by (N - 1), which significantly
improves performance on large range operations.

Run-All-Tests: true
Test: core-pager-writeback
Test: core-pager
Test: core-stream

Original-Fixed: 109983
Original-Bug: 102657
Original-Bug: 82818

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708011
Original-Revision: 7bff4c20e555698cafeb013563434b6f5e9a3b51
GitOrigin-RevId: 110a85cdbbeff1cee8e4724374384eb0917b2476
Change-Id: Idb70791339120f5bfe3d055d0821a2d3848bbb39