[roll] Roll fuchsia [bt][gap] Start LowEnergyConnector after insertion into container
Wait to start the LowEnergyConnector for new connections until after
the connector has been inserted into the container that the result
handler expects to find it in. This fixes a crash where an error is
reported synchronously during LowEnergyConnector construction, and the
result handler fails to find the connector in the container.

Original-Fixed: 116708
Test: fx test bt-host-gap-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771484
Original-Revision: 479d7654984e72134df8806c3a480fb80e6590f4
GitOrigin-RevId: f65e41f15ca55d9a48491348bc8b5ddb7bfea071
Change-Id: I2faffee3d940bea88364efc8b72930e24f82db34