[roll] Roll fuchsia [fs_management] Allow multiple guids/formats for PartitionMatcher
In cases where we might want to match against either a legacy GUID or a
newer GUID, and we don't know which one ahead of time, it is helpful to
extend the PartitionMatcher to take in a vector of type and instance
guids. Same goes for formats. This change allows for that flexibility.

Testing: fx test open_partition_test-component
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773673
Original-Revision: f8145cb04d4aaf18d6ab6839d6a9142e15cca656
GitOrigin-RevId: cb2241e748733f5b5c6bf56927fac7fa0f683b5d
Change-Id: I5c150b16439b6791eefb7251ef211c0a1e79eacb