[roll] Roll fuchsia [ffx] Build ffx-scrutiny exe and use in builds
This makes the scrutiny binary build in parallel as a separate binary
that can be invoked from the build, and modifies all the places we
currently have it doing so through ffx to use it.

This is not exactly how I wanted to solve this, but it works for now and
a better rearchitecting of how these components interact is for another
commit. It turns out to be annoyingly hard to improve these bits.

A weird side effect of the way I did this (generating fho handlers for
suites in the jinja template) is that ffx has to be labelled a subcommand.
This doesn't hurt anything, it's just an extra impl on the struct, but
it saved trying to hack in a way to tell if the suite was a subcommand
suite or not.

Original-Fixed: 116716
Original-Bug: 116712
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771667
Original-Revision: b4bf0e6b0c4af30ea860521f8e7e69d3ba12dbbb
GitOrigin-RevId: 05da0e8e355ec0527c13ff501cede5309260002f
Change-Id: I37cb9c7e1515241df1141b6198947f5d64f19f57