[roll] Roll fuchsia [dfv2][compat] UnbindAndRelease waits for Init
It is an error to call Unbind on a driver before it has finished
initializing, so move this check into UnbindAndRelease so it is
not possible to do this incorrectly.

Also make the root device of the compat driver have the correct init
state when being created.

Original-Bug: 117387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776163
Original-Revision: 9b645ceed6b1c879d79acb99e206357d227bc7b9
GitOrigin-RevId: dfc5fba0836f17dc7f45eb40d6af05f06eab7517
Change-Id: Ic564cd0fd004bae66229c532521f7f563ceb99df