[roll] Roll fuchsia [dfv2][compat] Unbind removes children
Have the UnbindOp function fully remove each of its children before it
unbinds the current device.

This goes through the proper flow of shutting down the NodeController,
Unbinding the child, and calling Release on the child.

This makes UnbindOp a function that returns a promise, instead of a
function that accepts a callback.

Original-Bug: 117387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776164
Original-Revision: 078998ccf6c4bb08e63c05e182950d3318cfbbbe
GitOrigin-RevId: cf9cde25c4db02e78fa6d7ac571da88565ce6b97
Change-Id: I6b807c82257e0491d5ad1ed417a2f25a8a49aec7