[roll] Roll fuchsia [driver_manager] Add `uses_fragment_driver` flag to composite fragments
This flag replaces a number of disparate uses of
`has_outgoing_directory()` which in the future will no longer be
synonymous with 'does not use the fragment driver'. This new flag is
more explicit and centralizes the logic in one place.

Original-Bug: 116754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777044
Original-Revision: c461e51120a410a20f59cd925345a80490482b4e
GitOrigin-RevId: 980a638cad149f223646634f8b3810a062a5a5f9
Change-Id: I838ebe5ad289585d334eeed1b4e153eca1d5a0e2