[roll] Roll fuchsia [input] Log keymap changes
We are getting a steady trickle of reports of spurious
keymap changes.

It is self-evident that this shouldn't be happening.

Since we don't have a good idea as to how that comes
about, I'm upgrading the debug log line for keymap ID
updates, so we can at least understand when this is
likely to happen.

This log like should be printed only rarely: when
the input pipeline first asks for the keymap, and then
on updates only. Since there is a singleton input
pipeline this should mean that the log lines will
be rare.

Original-Bug: 116681
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776930
Original-Revision: 4331119e7b25e7af3891d9d1653fdaaaedd5e693
GitOrigin-RevId: bbc23aa1a5759df620c1b86de7364f0a5f98c4d2
Change-Id: I0bbc0df7459a68907771e825a082637bf651e444