[roll] Roll fuchsia [cm][config] Introduce AbiRevisionPolicy
This CL introduces `AbiRevisionPolicy` to Component Manager's
`RuntimeConfig`. This policy contains the enforcement directives for the
ABI revisions found in `fuchsia.component.resolution.Component`.

The default is set to `AbiRevisionPolicy::AllowAll`, so that any call to
`check_compatibility` will return Ok(()) regardless if an ABI revision
is set or is compatible.

This policy is not yet used by Component Manager.

Original-Bug: 95188
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745064
Original-Revision: 1e047ce775f5ddf135101be181c94700deda719a
GitOrigin-RevId: 0ae9326be4e278fc13c198e2cbb9a47b1e0c2a16
Change-Id: I2d94e2f424d692dac36acad01ff2659bfe88e206