[roll] Roll fuchsia [fs_management] control component launch type
fshost has a collection and several static filesystem children,
depending on configuration. It needs to be able to specify whether to
use the static child or launch the filesystem instance in the
collection, since the static children are only used for specific
instances.

Add a ComponentType configuration enum which lets users specify which
type of child to launch. Make the default to be dynamic launching in a
collection named fs-collection, which is what most users want.

Test: fx test fs-management-rust-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773685
Original-Revision: 5db9e71d4b355c796d11d38cd7061ee863db4a2f
GitOrigin-RevId: 814c58642846f0c35e32923e89df3e5df9d3d84d
Change-Id: Icf490895b72fc65ee70758946edfc20bd0ceb5cd