[roll] Roll fuchsia [audio] fuchsia_audio_device::ControlCreator
This CL implements fuchsia_audio_device::ControlCreator. It
also adds Device class support for dynamic changes to gain
and plug states as well as the SetGain method.

This CL also introduces the concept that a `const` view of
a Device will be the limited subset of commands available to
device observers. As a temporary measure (until a future CL
that introduces the ControlNotify interface), the method
Device::SetControl is used to reserve a device so that it
can be controlled only by that client. SetGain is an example
of a method that can only be called when the device is
controlled.

Original-Bug: 87650

Multiply: audio_services_device_registry_unittests
Multiply: audio_services_device_registry_warning_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745606
Original-Revision: a168df05cd2b03bfce8c5997a200c62328e53bf7
GitOrigin-RevId: f15f562826e3e107e600f54c0ab90a2f57b7dbd4
Change-Id: I946dd4a6043c512ab3e7bceece0ff53586cab2ef