[roll] Roll fuchsia [ffx][emu] Move host tool hack to the only user
The issue referenced in the get_host_tool() function has been closed for
a very long time, but it did not completely resolve the issue (see
fxr/774028 and its revert fxr/776022) because ffx emu tests out of tree
still rely on this behaviour.

Rather than leave it in or remove it outright, move this over to ffx emu
where it's still used and tie it to a still-current related issue.

Original-Bug: 99321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776329
Original-Revision: c45793b01ec97cd333b732231f7990ea88ade907
GitOrigin-RevId: 8c5a01ae5bae9fbec43a6e326ba5955f5159feea
Change-Id: If46f280144d5908ec6cc570ac7528fce2224ec4b