[roll] Roll fuchsia [fidl][fix] Add ParsedTransformer
The |ParsedTransformer| is a base class intended to be extended to
perform transformations on parsed (but not yet compiled!) FIDL files.

There are some hacks and cut corners in this implementation (I've tried
to mark them with `TODO(fxbug.dev/114357)` as exhaustively as possible)
that we can optimize out as necessary, though my general working
assumption has been that for a tool of this kind performance is not the
foremost concern.

There is a good deal of subtlety in this |Transformer| class. As
discussed elsewhere, this is downstream of the fact that the current raw
AST design maintains two "views" of the source: the tree holding the
tokens relevant to raw AST node boundaries, and a separate token list
storing all tokens (including comments). Perhaps, in some future, an
alternative (or companion) parser can be added that stores all of this
information in a single data structure, thereby allowing a lot of this
code, which is primarily concerned with keeping the two structures in
sync as they're mutated, to go away.

Test: fx test fidlc-test
Original-Bug: 114357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767364
Original-Revision: ef50b1e3214e5093829accf1b9f03114c653215a
GitOrigin-RevId: 9c79e2dd4105f35f464bfb4e6ec7e6e0ba503d2e
Change-Id: I8d04dcbf1da08cc14c0648083c128322f012c4d5