[roll] Roll fuchsia [fidl][rust] Large message decode support
This implements the "happy path" for the large message decode support in
Rust. It does so by teaching the bindings to recognize both the
prototype (no `LargeMessageInfo` struct) and final RFC (with the struct)
designs. Once the encode logic is updated to only encode the RFC way,
the support for the old decoding method will be removed, and "Bad" sad
path tests will be added as well.

Test: fx test server-suite-rust
Original-Bug: 114259
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758953
Original-Revision: 888e0a50395fe15376d5505800674881debb9ca5
GitOrigin-RevId: 3c0addeed874198a32301258b9b724598778b397
Change-Id: I21c01016ea525c68fc790fb5b7b79c6d23d4e180