[roll] Roll fuchsia [ffx][product] use a single client for downloading
The client factory and its clients cache the access token. When
a new factory + client are made the access token is unknown, which
causes a request for a new token. In some environments the new
access tokens are rate limited. This Cl uses a single client for
all the actions in a single ffx invocation which not only helps
in those environments, but accelerates the download for all users.

Original-Bug: 103331
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775986
Original-Revision: 89524ab9bd5846176abd1d1df0a0a153d5c6e2d1
GitOrigin-RevId: 37b6efd0a87e57fa24bd9537b03e9350feb9ab94
Change-Id: I045f3033030a70cdbefe71210ce79a8954b48376