[roll] Roll fuchsia [identity] Reconnect accountctl.
Accountctl is a small tool we wrote to help debug the account system.
We haven't used it since CFv1 and somewhere along the way it must have
broken. This CL reattaches it using knowledge of the account manager
location in the topology.

Original-Bug: 117217
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777662
Original-Revision: 21225e130052357b31352ef8c6f09e464f0a3fae
GitOrigin-RevId: 12a0426b433049abdee0de64047a84395b4b9f5e
Change-Id: Iccbb793152b91416e9915c65e18269e073976e8d