[roll] Roll fuchsia [ffx][repo] Start should err if server is running on unexpected port
This changes `ffx repository server start` to error out if after we
started it, the server is running on a port we didn't expect. This can
happen if someone changes the listen port in the config without
restarting the server.

Original-Fixed: 117737
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778769
Original-Revision: b24ed273c815e7271b22a893b83d799e39724cfc
GitOrigin-RevId: f89255529b56bed0b4fe0b3567092614acef47ae
Change-Id: I362269b4756635b0d65778e7104f6dd6935a1a65