[roll] Roll fuchsia [audio] Adjust Pinecrest output driver valid_bits
AudioCore is currently limited to the 24-in-32 sample
format, if the container is int32. This change adjusts
the Pinecrest driver from 32-in-32 to 24-in-32. This
doesn't change anything about the signal, it just
communicates a format that AudioCore will accept.

The in-progress refactor of AudioCore will address this
AudioCore limitation, and fxbug.dev/117743 tracks the
restoring of this driver to full AUDIO_SAMPLE_FORMAT_32BIT
once that V2 audio refactor is landed.

Test: signal_generator on Pinecrest
Original-Bug:  110904
Original-Bug:  117743

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778306
Original-Revision: 853768e5e97b049c74723c5ed12543fb44570c2c
GitOrigin-RevId: 3f7aef04885aa5cf9feb6bc90c739d67d4bcbae9
Change-Id: Id5c18b07560d44d9065899b9d2379c615b466220