[roll] Roll fuchsia [netstack3] Allow listen backlog <= 0
Implement the POSIX requirement that calling listen with a negative
backlog is the same as calling it with a backlog of 0, and implement the
allowed behavior of imposing a minimum backlog size. While imposing a
minimum, also impose a maximum of 4096 to match Linux behavior.

Original-Bug: 117163
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773145
Original-Revision: 58f80778fc793b61bd2f0020541f3ae2bb94b8c4
GitOrigin-RevId: 6b6b3491a437f6340a7a4cd15359d5f25f13051a
Change-Id: I6719f407cb1f5a54a24d5e9de492d252418e2d73