[roll] Roll fuchsia [netstack3] Allow peek offset past ring buffer
The SendBufferWithZirconSocket reports as its readable bytes the sum of
the readable bytes in the ring buffer and the bytes in the zx socket.
That means a user of the type could reasonably try to "peek" at an
offset past the end of the ring buffer. Handle this case instead of
crashing by providing an empty slice to peek from.

Original-Bug: 117259
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778742
Original-Revision: 739427d5f916321dbfb0d6276073dc22ae646320
GitOrigin-RevId: 6868c196b4469149ede17e85a24824f278c8c7f5
Change-Id: Iacf722ab67ddc63ff908b78c3a73d2e2e3a3fc5a