[roll] Roll fuchsia [build][bazel] Fix platform_aibs() for Bazel 6 (again)
This is a reland of [1] which was reverted because while it fixed
an issue for Bazel 6, it introduced another one for Bazel 5.x.

This version hopefully should work for any version, by adding a
new utility function that deals specifically with differences in
behavior between Bazel 5 and 6.

[1] https://fuchsia-review.git.corp.google.com/c/fuchsia/+/772744

Original-Bug: 112021
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778822
Original-Revision: d77fdf8099b24d15d9c2cdb2e8b9fe4b625d3d03
GitOrigin-RevId: f5c269cbffb5e532a721ed5242acd1d45b4aec6b
Change-Id: Ibf1e206eff563225c03346cdc0ba272be48727a1