[roll] Roll fuchsia [fidl][fidlc] Remove outdated visitor_unittest
This test previously served two functions: ensure that all visitors were
called, and ensure that the file could be pretty printed using this
method. The first case is covered well by `span_tests.cc`, as well as
numerous other classes that inherit from the |TreeVisitor|, like the
linter and transformer.

The second case is still technically possible, but relies on code that
is being removed. Since the new canonical "pretty print stuff using the
tree visitor" tool is the formatter which builds on top of the tree
visitation functionality, this test becomes redundant. On top of that,
it is subtle and hard to maintain.

Test: fx test fidlc-test
Original-Bug: 117642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777928
Original-Revision: 590b7cdf92f2afa860eb285b850e0888e55f4d7b
GitOrigin-RevId: f3fa41b207fdf32493308725ea24fde2a6b0ad60
Change-Id: Ib88090e21c86914cae8bbc6577d189347214e4d8