[roll] Roll fuchsia [driver_manager][DevfsExporter] Reduce repetition
Extract DevfsExporter::Export that can be shared by Export and
ExportOptions. Store exporters in a map so that they can be removed in
O(1) when devices are removed. Pass the watcher to its on_close_callback
to avoid the need to capture the pointer.

It would be even better to change `DevfsExporter::exports_` to
`std::unordered_set<ExportWatcher>` but that would require constructing
the `ExportWatcher` in-place, which would require more refactoring.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778962
Original-Revision: 39b42fdb013691173e65b4973843e37a834c223b
GitOrigin-RevId: 058750413b500e081691e92d797565dc39a03e37
Change-Id: I8ea41ab5420515603ef83e4d3eb8b0889c05b6e6