[roll] Roll fuchsia [netstack] Flush caches in FIDL mutators
Fast UDP sockets maintain multiple kinds of distributed state:
 - Payloads are enqueued by the client into a zircon socket and wait
   there until they are dequeued by Netstack.
 - The client caches metadata locally in order to avoid having to
   validate it over FIDL on each I/O.

When the socket is mutated, we have to take care to ensure that this
state is handled correctly:
 - Payloads currently enqeueued into the zircon socket should be
   processed with the socket state before mutation.
 - Client caches should be flushed so that validation can happen with
   the socket state after mutation.

Before this change, we have already been performing these operations on
FIDL setters known to affect socket state in relevant ways. Here, we
expand this coverage to be performed on all FIDL setters. The reason is
that we are soon going to turn on Fast UDP and would like to minimize
the surface of possible errors (much of which comes from the surface of
socket FIDLs). Once we've established confidence in the Fast UDP
protocol in production, we will pare back the set of FIDLs for which we
flush caches.

Additionally, this implementation requires that new FIDL methods in
`fuchsia.posix.socket` be explicitly implemented on `datagramSocketImpl`
structs. The goal is to force future developers to consider whether a
cache flush is necessary.

Original-Bug: 109784
Original-Fixed: 106636

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726842
Original-Revision: b26bcb6a7465053f30b5bd56e972693b441fbb24
GitOrigin-RevId: 866a82cf0442ead61cd3d291f9178c6d714df9e3
Change-Id: Iff4c01489fa890ab4a01ca877af9321b3736f7b8