[roll] Roll fuchsia [zx] Format kazoo goldens and pause golden diffing
We temporarily dismantle the golden diffing of kazoo outputs and format
what is checked in. This will make it easier to reintroduce the diffing
with the formatted Zither outputs and clearly see the difference.
(kazoo outputs are not formatted by default.)

Original-Bug: 110295
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774804
Original-Revision: f0b8be17248f8282cc1af1850d1143c300a3bd65
GitOrigin-RevId: 64b41296f23c818bf7240795e19127ed681f38e6
Change-Id: I1aa50644f0ddaae9244c353d5d86e802eb304fd2