[roll] Roll third_party/go [zx] Migrate generation from kazoo to zither
See fxrev.dev/774805 for the fuchsia.git change that has kazoo pass the
baton to zither, a new FIDL backend.

The notable diff here is due to addressing a previous inconsistency
in the generation kazoo's 'syscall/zx' syscall wrappers: sometimes
more conventional type bindings would appear in the signatures;
sometimes the more 'FFI-friendly' type analogues would (e.g.,
`unsafe.Pointer` vs `*MyStruct`). Since the intent seemed more to use
conventional types in this case, especially those defined within
'syscall/zx' alongside the bindings (e.g., `Handle`), this change
uniformizes things in that direction. As a result, we have to both
introduce several more types to the package - as well as tweak the
current hand-rolled definition of `PortPacket` so as not to be an
interface.

One of the main goals of fxbug.dev/110021 is to get to a place in
which all syscall data types like `PortPacket` are generated directly
from FIDL library zx. As a step toward that, we make this type, among a
handful of others, look more like something that could be
machine-generated. At this point, outside of the port packet case, this
mostly just meant renaming a few choice elements so as the CamelCased
versions of the associated C type names.

Original-Bug: 110295
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/778232
Original-Revision: ad59569b45b5941fb8431f2fce6afa86bc72f6cd
GitOrigin-RevId: eb8abadc9c3a20854f805e17f8fc97846f72cd17
Change-Id: I4b62dbc0317e42e7d3f9eca9d7a5a61811880ad5