[roll] Roll fuchsia [starnix] Clarify UnixSocket.
I found the query_events() code to be very confusing. This adds a
comment explaining how the data is stored.

Moves the implementation of UnixSocketInner.query_events() inline to
the only caller, UnixSocket.query_events(). There are three query_events
functions involved (UnixSocket, UnixSocketInner, and MessageQueue) which
each return slightly different things. The "inner" version returned only
the part of the answer that could be computed without locking the peer.
I found this subset of behavior hard to reason about and hard to
describe in a comment, so moved it inline so there is one computation in
one place for the whole thing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778986
Original-Revision: 176d09f10268a777ddc48d187202e515474a0ca8
GitOrigin-RevId: fb289c63976033015bf8a9d4f40a186c2f48ad80
Change-Id: I494531d3d7d0df9745b58963b32c4cf2f98fcd7b