[roll] Roll fuchsia [zx] Replace kazoo usage with zither
This change replaces all in-tree usage of kazoo with zither,
specifically to generate the "cdecls" backing the public headers, as
well the kernel/vDSO and rust syscall bindings of library zx.

As evidence of this change being a 'no-op', we replace the kazoo
goldens with their zither counterparts to reflect the absence of any
real difference in the generated sources.

One notable diff here is due to the go_runtime zither backend addressing
a previous inconsistency in the generation kazoo's 'syscall/zx' syscall
wrappers: sometimes more conventional type bindings would appear in the
signatures; sometimes the more 'FFI-friendly' type analogues would
(e.g., `unsafe.Pointer` vs `*MyStruct`). Since the intent seemed more to
use conventional types in this case, especially those defined within
'syscall/zx' alongside the bindings (e.g., `Handle`), we opted to
uniformize things in that direction. In any case, only one in-use
syscall signature was changed - replacing an *int with a *PortPacket
(same alignment) - though the difference should be a no-op and has been
approved in fxrev.dev/778232. Also, we fix a bug in which the
zx_string_view_t was treated as an object of size 8, i.e., simply as a
pointer; however, this syscall is unused and so there should be nothing
affected.

A follow-up change will delete kazoo outright.

Original-Bug: 110295
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774805
Original-Revision: 2c9912486c184ab4e21b5578502b45d9238fe1e9
GitOrigin-RevId: e95497803245ea035331dc0fbbc93597e0be0670
Change-Id: I1c3f1400f5ac5804b48333ccd07cbb5d50ab6308