[roll] Roll fuchsia [kernel][port] Improve scalability of ports and Semaphore
The purpose of this change is to reduce thread lock contention caused
by ports.

PortDispatcher uses a Semaphore to wake threads that are blocked in
zx_port_wait.  Every packet generated results in a call to
Semaphore::Post and every call to zx_port_wait results in a call to
Semaphore::Wait.  Prior to this change, every Wait and Post operation
on Semaphore acquired the thread lock regardless of whether there was
any need to block (in the case of Wait) or any waiters to wake (in the
cast Post).

This change improves port scalability by eliminating the need to need
to acquire the thread lock in the following scenarios:

* zx_port_wait when the packet queue contains one or more packets.
* Queuing a packet when the port has no waiters.

Original-Bug: 111882
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730292
Original-Revision: fc0dea02a8e2b71a49b296f18ee59dd784cff3d4
GitOrigin-RevId: 887177b4e9d8bbad45505988635c9912ae1468f1
Change-Id: I784e42d3baea3525ea305c7cf84bf41b9d0bc742