[roll] Roll fuchsia [starnix] Add tests for the MAP_GROWSDOWN / PROT_GROWSDOWN feature
This adds a set of tests for the {MAP,PROT}_GROWSDOWN feature commonly
used to manage stacks. The tests cover growing a region and the
interaction with mappings in or near the guard region. This feature is
unusual in that it changes the effective protection of virtual addresses
that are not contained within any mapping the tests use a combination of
memory accesses and MAP_FIXED_NOREPLACE mappings to probe the effective
state of the address space. PROT_GROWSDOWN is also notable in that it
applies to regions outside the specified area and can modify the future
protections of unmapped regions.

Original-Bug: 97514
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778983
Original-Revision: e8c220dbcef871107c53ddf50d30de41857074d1
GitOrigin-RevId: 04b884f2a6e190c4f6c4a28dd54c20a01e131d79
Change-Id: Ia479ecf6d5c2330c1ea677db63eb5c2c63b52e12