[roll] Roll fuchsia [starnix][syncio] Pin zxio storage
It is possible for zxio objects to be self-referential; for example, see
the zxio_datagram_socket_t's RouteCache. This means that code that
manages that underlying storage must ensure it is not moved, as that
would invalidate existing references.

Document this invariant in zxio, and enforce it in syncio by pinning the
storage for the zxio object on the heap before calling into any
initialization logic. Expose a safe API wrapping this logic to clients
such as starnix.

Test: fx test starnix_gvisor_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778753
Original-Revision: bc47f52ff4f1e33f845840661a294586562f35b7
GitOrigin-RevId: fbf8489c9e634e1cffbef22d2297766c614297bf
Change-Id: I31ba182b91ac237bdfda08bb7079aab781490bfd