[roll] Roll fuchsia [fidl][rust] Enable large message encoding
This change enables large message encoding in Rust, and adds the
relevant dynsuite tests. It additionally does a bit of cleanup in
existing large_message_tests.cc code.

While this feature is "enabled" in the sense that it now exists and
works according to the spec laid out in RFC-0196, it is still behind a
fidlc feature flag (`allow_overflowing`) and requires explicit
annotation of each method where it is used.

Test: fx test server-suite-rust
Original-Bug: 114259
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758954
Original-Revision: 91898dac5ed01659e8fd798550ba5becdb5320da
GitOrigin-RevId: fb4395474563a622492d009b70b8e41a1bf7dd4e
Change-Id: Id2a65c0f7c616f92caefef4049045b885be4eb44