[roll] Roll fuchsia [outgoing] Add |AddProtocol| unique_ptr overload
The |AddProtocol| overloads that take a raw pointer to
|fidl::Server<T>| or |fidl::WireServer<T>| are unsafe in
that they require users to coordinate the lifetime of the
servers with the |OutgoingDirectory| object. This new overload,
takes a |std::unique_ptr<ServerImpl>| which is safer because
the |OutgoingDirectory| owns and can ensure that the server is
destroyed and active connections are closed before going out of scope.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/775483
Original-Revision: 0152f245b7d7208ee15a567a8fc7d5bb1881a2f3
GitOrigin-RevId: fb99320c88ca5a9c68857d9700197f114f19cdfc
Change-Id: Ia720cdb0c04f2ba13ec40b8c1e8f4ab5ae89da1a