[roll] Roll fuchsia [xhci] Remove TRBPromise when TRB is unused
TRBPromise is often used throughout the driver for convenience, but the
TRB contained in the promise is unused and unnecessary. This makes it
harder to understand what the code is doing. In the interests of
optimising for the reader instead of the writer, this CL removes those
unnecessary uses of TRBPromise and replaces them with promise<void,
zx_status_t>.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765564
Original-Revision: 86026d4bacf290c861f8f6812e2f6b6f44b3e30c
GitOrigin-RevId: b8bc8a3875c911f5805ca4f1551f96c57a21b803
Change-Id: Iaae623d31b93ab48bcc83734cbde1bd415a31d5f