[roll] Roll fuchsia Reland "[flatland] Remove BufferCollectionImportMode"
With prunable tokens implemented in DisplayCompositor we're down to
two buffer collection import modes: attempt to use display composition
or don't.
This CL removes the BufferCollectionImportMode enum and replaces it with
a boolean.
Since this reverses the order in which the build flag and the config are
applied, we rename the flags to set ensure the default remains "display
composition off". Doing this should avoid triggering fxbug.dev/117785

Original-Bug: 117690
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779702
Original-Revision: a0fd080606940836a23e41c4dcba206bd389bd31
GitOrigin-RevId: 448e29d83884e58ee547087873c447165b636782
Change-Id: I0e980af1657fefe71251f8a37aa0d4ee1d23b3b8