[roll] Roll fuchsia [ffx] Make ffx target an explicit ABI revision from API level 10.
Targeting the latest ABI revision doesn't work well in practice as it
means that we will release a version of ffx that isn't compatible with
any previous version of the platform, and so users won't be able to use
this version of ffx to update their device. Targeting an historical ABI
isn't technically correct but it should provide a "grace period".

This is a temporary fix to unblock some users, a cleaner solution will
be provided shortly after this.

Original-Bug: 117872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779864
Original-Revision: e38ded2eda651e11d08eabcd9572c6bab5fbd60e
GitOrigin-RevId: e0c9880422a9d13aadf4915257b639210f9b191b
Change-Id: Ie6a99605eac39e20b4458b0978e5b3b40bca8187