[roll] Roll fuchsia [netcfg] Have dhcpv6::start_client return stream
... rather than storing the stream into the StreamMap itself. Also
don't attach any context to the error on the stream to avoid double
printing it (the relevant context is readily available when an error
is yielded by the stream and already printed).

No change in behavior; this is a refactor to prepare for upcoming
changes where we also need to return a stream as a result of calling
`WatchPrefix` repeatedly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762175
Original-Revision: 3fcc06774554fd7a0e5d0d47c9bd542cada5e4c7
GitOrigin-RevId: 5d885501bc82cce1b35f9ac3392036e33dceabd6
Change-Id: Id2c7e430c383186fa2076b6e4a230f629c9af186