[roll] Roll fuchsia [gazelle] Improve readability of shortcuts image.
The text on the post-it note isn't very legible
and there are some artifacts on the image. I made the
font larger and more bold, and got rid of the rotation
because it was creating some artifacts.

I'm assuming this is placeholder and so small tweaks
don't need to be run by UX.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780062
Original-Revision: 823a508ce9e2ee99d71cb0e45586af2bb149e47c
GitOrigin-RevId: 1b4b732ea4bb1cb8a53da4d6de998db7fe63fbd0
Change-Id: Ie5f77d020be971f92f8d81a0f469e4bda4a7a3b7