[roll] Roll fuchsia [netstack3] Prevent overriding addr zone for TCP
Disallow setting the bound device for a socket if the socket already had
a device set and is bound or connected to an address that requires a
zone. This ensures that, once support for zoned addresses is added, any
zone set via binding or connecting can't be overridden by setting a
device on the socket.

Original-Bug: 115524
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779944
Original-Revision: 00ab699049d93507239e9b2272f9b828347306c0
GitOrigin-RevId: 99eeff67f7579c189b086ee65230bef85d8972d3
Change-Id: I82c612bbf24e3b00df12ad37b54c20156ef19a32