[roll] Roll fuchsia [outgoing] Rename AddProtocol to AddUnmanagedProtocol
To better distinguish between the managed overloads in the
|OutgoingDirectory| class, and the ones unmanaged, the latter
kind have been renamed to |AddUnmanagedProtocol| and
|AddUnmanagedProtocolAt|. The managed overloads take a pointer
to a protocol server and closes the connections during
|OutgoingDirectory| destruction or |RemoveProtocol|. The unmanaged
ones on the other hand do no such thing. Having a distinct name
better signals to users which is best which use case.

Original-Bug: 89036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780107
Original-Revision: 38107164bd572c831350443e236492e0f3375ba9
GitOrigin-RevId: 0c9c1db3b1bd34c5928e2818c5ece2e92be20cd6
Change-Id: If2106bf207402d446db651ab4309bbdb54e52c56