[roll] Roll fuchsia [network/tests] Explicitly depend on sync UDP
Introduce a Sync UDP variant of the netstack and depend on it explicitly
in the set of tests that need to run against both Fast UDP and Sync UDP.
This lets us flip the default in a child CL in an easily revertible way.

Once Fast UDP is turned on, we will remove the explicit Fast UDP
components, since they will then be redundant with the default.
Depending on whether we think we need the coverage, we might also remove
the Sync UDP components at some point.

Original-Bug: 117436, 117883
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779671
Original-Revision: 396c723549316fe0bdd24ea6bb38466f4a1f4164
GitOrigin-RevId: 82d18b3458e4ac3245eb2929d776e1dfa901271a
Change-Id: I63ceb9372d953a16fae5f9aa8ef8ec832be638bf