[roll] Roll fuchsia [doc-checker] Fixing link to old markdown file
The presubmit program `doc-checker` does not allow links to files in
//docs that are not at HEAD of the main branch. In this case, the
document intends to reference the older revision explicitly, so the
preferred way to implement this is to use a link to cs.opensource.google
which is allowed by doc-checker. Another alternative to consider is
to link to the Gerrit change which modified the file, if that is more
appropriate.

Original-Bug: 117653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780091
Original-Revision: 9caeea3967f737f34b994cf3d3933de703ac6c30
GitOrigin-RevId: ccd2e5bd39d8be602a964bcab0af132cabf48900
Change-Id: I35414e59135076e9b5ad7d8e10c0d31681fc6e03