[roll] Roll fuchsia [doc-checker] Fixing references to topaz dart
This fixes links that point to the removed project "topaz" and
to in-tree Dart build rule files that have been moved.

The exact contents of the Dart build configuration is most likely
out of date, but is outside the scope of this change.

Original-Bug: 117653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780622
Original-Revision: 83c62dbda64594f97d0687c2afb912a4e6f1e4af
GitOrigin-RevId: 5529cc9aec81f5cc0666583b39d0e300fc5f179a
Change-Id: Ibf8c3a3f20acfb4f15df2424b173b9a9bc5c09b8