[roll] Roll fuchsia [doc-checker] Fixing  ambiguous links in md fragments
The presubmit program `doc-checker` checks links to local documentation
using filepaths. To enable this, links to local files should use file
paths. This is complicated by the use of included markdown fragments.
Normal practice is to place these fragments in a `_common` directory
and use `<< >>` to included them in a markdown file. When published,
this inclusion is done before links are rendered, so the links in the
included file are relative to the "host" file of the fragment. This
breaks the file path checking in `doc-checker` so full file paths are
preferred.

Original-Bug: 117653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780623
Original-Revision: 5321ca22da16521f5f8e3a15348b0514b1a70f20
GitOrigin-RevId: e2aa8c6ce9d1f896b65dea156ab70285a646133b
Change-Id: Idb59446a8d514c409ad2adfb6e2db778f9721b58