[roll] Roll fuchsia [check-licenses] Fix SPDX ID creation and references.
Previously, check-licenses would try to guess what the SPDX ID for a
given license file was by using the path to that license file.

This _could_ have worked, but it was overly complicated. It was
implemented incorrectly, so the license texts were using a different ID
in the SPDX document than expected --  which causes issues when
downstream customers attempt to parse the resulting json file.

This CL updates the code to use a simple auto-incrementing number as the
IDs, and references to these IDs use the object variable directly
instead of attempting to re-generate the ID.

Tests will be added in a future CL.

Original-Bug: 117959
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780645
Original-Revision: 82e34c7cfc5458ca60ff372d4250d9e6fb35664f
GitOrigin-RevId: 1a94976ef565207abfafcf368e0c3203393aa4a6
Change-Id: I32074bf4a90e53da695a89cea0398a00c9bd0f99