[roll] Roll fuchsia [events] Hooks::dispatch cannot fail
Currently, hooks are allowed to fail, which in turn can fail
the action that dispatched the event.

This should not be allowed. Hooks should not be able to affect
control flow. They are a unidirectional notification system
of important changes that occur to components.

If a hook produces an error, a warning message is printed out
by component manager, but control flow is not affected.

There are two tests that relied on a hook's ability to change
control flow: `destroy_error` and `shutdown_error`.

To work around this, we have introduced the ability to mock
results of actions in `ActionSet`. Instead of creating hooks
that fail an action, we now mock a failure for an action
instead.

Original-Bug: 116855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778752
Original-Revision: 108abfa498635ec999f90fc948ec8c4fee89fe1a
GitOrigin-RevId: faf174b0de0277c5b9bded36940efd1bfd6ade90
Change-Id: Ic88f25bf73c7e2c9f5b9309815a9684a16271483