[roll] Roll fuchsia [assembly] Move component_manager_bootfs_config out of common
This config is not used in any real products, therefore it should not be
part of the common set of platform AIBs. Each product that wants it
should add it explicitly in their product gni definition until there is
a proper assembly feature to enable it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780863
Original-Revision: 597e33f7d41eb5f2daf00a9f19519cc377c35f10
GitOrigin-RevId: 50f0e48cd71cda0c16464054c764da564b45a7b7
Change-Id: Id051c3108353578386884336592873def773fa6d