[roll] Roll fuchsia [recovery] Add ShredDataVolume to fshost
This moves some of the responsibility for shredding the data volume into
fshost and fixes how shredding works when the data volume is mounted:
instead of trying to wipe the super blocks, we will delete the key bag.
For now, this only works with Fxfs and only in the Rust fshost.

This CL does not fix workstation where factory reset remains broken.

Original-Bug: 116488
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780203
Original-Revision: 732c98172635005437819cc5e3c54bdf148a6ace
GitOrigin-RevId: 3f63bdf965988aa3adf8bf72814c1d1e755fa10f
Change-Id: Ifd382db548e2fb2d2868600750d5ba217ac0b9e1