[roll] Roll fuchsia [wlan][policy] Don't connect without interfaces.
This change modifies IfaceManager such that it will not
attempt to connect if there are no client interfaces.  The
result is that policy API clients will be notified that the
device is in an incompatible mode if they request a
connection before starting client connections.

Original-Bug: 117485
Test: act.py \
    -c config.json \
    -tc StartStopClientConnectionsTest
Test: OOBE
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781702
Original-Revision: cbac4ca0d4c24bf154eb9505209eab24ffe1afd1
GitOrigin-RevId: 2286e62fc7990410dae6897ccd176affa2aa0f96
Change-Id: Ib3bbb0e8378e3d5dbc99608eb3348570d1dae3b7