[roll] Roll fuchsia [cm] Introduce ABI reading library for in-tree Component Resolvers
Introduce ABI revision reading API for Component Resolvers to read a u64
value from a provided file path in a directory. This function does not
error if an ABI revision is not available.

This function may be used by in-tree component resolvers to read an ABI
revision and if it exists, set the value in `fresolution::Component`.

Original-Bug: 95188
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754619
Original-Revision: 120be5caf00d49d0197836bf9bdb373f821a31ce
GitOrigin-RevId: 7a3de11837e2825d191466fb3def315d9318f3da
Change-Id: Ib226a9957f7c0236ed2caa893e68e83be2aa179f