[roll] Roll fuchsia [wlan][policy] Fix flaky read of FIDL channel
Several of our tests put a message into one end of a FIDL channel,
then try to read from the other end with `stream.next()`. Occasionally,
the first attempt to run the `stream.next()` future returns
`Poll::Pending`, as though the kernel hasn't realized the message is
in the channel yet.

We chatted a bunch in the Rust chatroom, but couldn't find the _real_
underlying explanation. However, because the message is really in the
channel, retrying the attempt to read it works. There's a bunch of ways
to retry, but exec.run_singlethreaded works reliably in my local
testing. We could also loop (like we do elsewhere in the code), but
that's less elegant.

Original-Fixed: 116134
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781465
Original-Revision: e9dc8ae45af2431d50d1eee596a0cea67634ef42
GitOrigin-RevId: 00ce0cf8e1290fff29ebc8203cf187242c78d327
Change-Id: I9ec3bbb5f39fb1370d602d0c4d3fe41be8a70a8b