[roll] Roll fuchsia [fidl][examples] Key value unknown interactions.
Add unknown interactions to the key-value-store examples. This changes
the main protocol to `open` and methods to `flexible`. This is mainly to
demonstrate that open/ajar exists rather than provide a specific example
of how to use them for protocol evolution; rather we just want them in
use since open/flexible is intended as the default in the future.

The add_iterator example does demonstrate a case where a protocol might
want to be declared as closed; the Iterator protocol is associated with
a single method which initiates iteration, so it is declared
closed/strict rather than open.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780003
Original-Revision: 3637024fa9af082593d88b41cc841d9ba6bb6a73
GitOrigin-RevId: d55a52d8199c68ce2c115f007070b4247577d3b4
Change-Id: Ia4b999b301bd9246e17f621a99a9466a3236a807