[roll] Roll fuchsia [assembly] Add logic for WLAN assembly
Add logic to allow assembling different sets of WLAN packages based on
(1) level of privacy support and (2) whether the device is SoftMAC or
FullMAC. This logic is gated behind a flag, which is not set to true in
any platform configuration yet.

This additionally helps to unblock WLAN's USME migration project, as
WLAN needs to include different packages for FullMAC and SoftMAC during
the migration, and assembly has more meta information available to make
this decision.

Original-Bug: 117249
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769011
Original-Revision: fe5148526937fda3abda291d30d5bc88f0964281
GitOrigin-RevId: c811ae75c4d584f82e41a9f69c836efe0f8734b7
Change-Id: If35fd22145d1dbc37e2c58eb12d018fee900b374