[roll] Roll fuchsia [audio] remove spurious command timeout
An unnecessary driver command timeout was being logged, when
moving the driver to the Start state. It does not seem that
this is the cause of the linked bug (since this timeout only
causes logging, not any change in behavior), but the log
message was found near the time of the problem.

BUG: b/262607079

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781662
Original-Revision: 04436807cc387a5aaae6b7f0b05805f1cf484148
GitOrigin-RevId: 0826a68cff1d3b22e853cc106ef432c89305bbb9
Change-Id: Ib6ddc855a2be4edbdfb33a156944c594a13a0b81