[roll] Roll fuchsia [build][sdk] Rename private_headers to sdk_headers_for_internal_use
"private_headers" lets the PlaSA pipeline skip over headers that can't
be directly included. Currently it doesn't do more than that. But the
name wrongly suggests that it may teach the build system to mark
headers as private. Here we rename it to be explicit that:

- These headers are SDK headers. They will be included in the
  SDK and may be included (indirectly) by user code. I avoided the term
  "public" because the definition of public in GN is different from the
  definition of public in the SDK.
- These headers are omitted from the PlaSA enumeration process.

This CL also aligns the behavior of sdk_* templates w.r.t to whether
sdk_headers_for_internal_use are merged into all_headers under
the hood. Because two templates do not perform the merge and one
template does perform the merge, I picked the option to not merge to
reduce changes. But it should be fairly easy to start merging them if we
decide so, in which case we can drop some entries from `public` in quite
a few GN targets in our tree, which will also improve the effectiveness
of `gn check`.

Original-Bug: 117085
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778626
Original-Revision: 4458f5c2c6e07603622e40b07ddce37fb1f7c327
GitOrigin-RevId: 6be23b129dfb806bd2e9a2ab27a1cc1daa25aab9
Change-Id: Idd976169ef9f6c6ab7ab0074d3ea8279f45852f9