[roll] Roll fuchsia [settings] Rename MessengerClient::fuse
`MessengerClient::fuse` is intended to hold a handle to cleanup code
that will be called well all references are drop. This simply renames
the variable to `_fuse` to document that it is intentionally unused and
only there for scoping.

Test: fx build
Original-Bug: 84729
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781706
Original-Revision: 42e048e2ec55a7d9b8feb819a428c534fa244319
GitOrigin-RevId: 016ff52b10382c35f9fc83f832e2d70d1ceb3e8d
Change-Id: If98a7be021d6577a028b8a5ac1057ecfa7f3d1d2