[roll] Roll fuchsia [audio][drivers][as370] Reduce reported DMA transfer size
The Dhub DMA engine copies data in MTUs (Maximum Transfer Unit)
and also waits for a previous MTU to near complete before starting a
new one. Hence, for audio output in a subsequent CL we can report
this MTU size as the main cause for internal_delay (calculated from
fifo_size in simple-audio) reducing the reported latency.
This does not change the rest of the output driver behavior, in other
words the latency for output is already lower than previously reported.

For audio input we keep reporting the internal_delay from the copy
of a number of MTUs (dma_mtus) and the SW PDM to PCM decoding time
since that is decoupled from the DMA transfer size. The input driver
does not use transfer size since it relies on the size_per_notification
value as that is how much data is DMA before the SW PDM to PCM decoding
starts. In other words, this change does not affect the behavior of the
input driver.

Original-Bug: 107645
Test: fx test -ov syn-dhub-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777987
Original-Revision: 32426e26fdd8174be49ee2a0fb45c3d82bb9a620
GitOrigin-RevId: 4f4680f64d0878f145d5c6d4461c62b5ed7195ef
Change-Id: I6cf5b51f91dbfa7c0f87869c6156af417901b537