[roll] Roll fuchsia [fastboot] Migrate to |fidl::ServerBindingGroup|
This CL migrates |AddProtocol| usages where a raw pointer
to a FIDL server was passed as a handler. That overload is
deprecated. It is deprecated because the method overload can
hide subtle lifetime bugs. Instead, servers should use the
|unique_ptr| overload which gives ownership of the server to
|component::OutgoingDirectory|.

This was not done in this CL. The use cases migrated here passed
a `this` pointer. Thus it was hard to migrate to a |std::unique_ptr|
based approach without major refactoring. Consequently, this change opted
to instead leverage |fidl::ServerBindingGroup| to own connections, and
using the generated handles to handle connections.

Original-Bug: 89036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780334
Original-Revision: a2e10a39d8962f8d75d73713b42bcee9dc31809a
GitOrigin-RevId: 31617b6ff5d5110dfc6baa60ba60cc6970eb8b66
Change-Id: I09921bd30baa99d998bdc444330072d00a56248d