[roll] Roll fuchsia [devices] Migrate to |fidl::ServerBindingGroup|
This CL migrates |AddProtocol| usages where a raw pointer
to a FIDL server was passed as a handler. That overload is
deprecated. It is deprecated because the method overload can
hide subtle lifetime bugs. Instead, servers should use the
|unique_ptr| overload which gives ownership of the server to
|component::OutgoingDirectory|.

This was not done in this CL. The use cases migrated here passed
a `this` pointer. Thus it was hard to migrate to a |std::unique_ptr|
based approach without major refactoring. Consequently, this change opted
to instead leverage |fidl::ServerBindingGroup| to own connections, and
using the generated handles to handle connections.

Original-Bug: 89036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780333
Original-Revision: b4d17c2591144a5147646a7b21ab735cda56c086
GitOrigin-RevId: 946988fe117ee4e2133d8dd8ce97410248475763
Change-Id: I06f2c01e4816b275d2a19fe15653cf644193c98a