[roll] Roll fuchsia [fidl][dart] Refactor message decoder
This change refactors the `decodeMessageWithCallback` function to be
more agnostic to its position in the message buffer. Previously, it
assumed that top-level members were always offset `kMessageHeaderSize`
into the buffer. For large messages, that assumption will no longer
hold, unless we want to decode the header twice (we don't).

Instead, we simply have the callback supply the offset as an argument
(as of this CL, always `kMessageHeaderSize`). This brings us closer to
the C++ and Rust decoder designs, which operate on the idea that they
are decoding two sequential FIDL messages (a struct followed by a
header, consecutively in the same buffer), rather than special-casing
the transactional message.

Test: fx test fidlgen-dart; fx test check-goldens
Original-Bug: 114261
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781609
Original-Revision: 99210b175d86d460d5c314d21d76e8c8d9bba5ee
GitOrigin-RevId: 1eeb2bf9dd192d42c851016006433584f967d623
Change-Id: I48687f9a5a859cfffb25f0b4d5803714b62f5daf