[roll] Roll fuchsia [fshost] Support unformatted zxcrypt in netboot
When netboot is set, WriteDataFile only works if the data partition
happens to already be formatted as zxcrypt and if the driver is bound by
something else (i.e. the paver).

This means that SSH keys are not persisted when paving if the data image
payload is absent.  This is silly, since we don't actually care about
the data image payload's contents during paving (and prevents
fxrev.dev/781242 from landing).

Note that this also required a fix to the paver: The paver will bind
zxcrypt during partition initialization if needed (i.e. if the partition
in question is marked as a zxcrypt partition), but it never unbinds the
driver, which prevents fshost from functioning.  I could have fixed this
in fshost by attempting first to attach to the exisiting zxcrypt
instance, but that would be inherently racy, so I chose to fix it in the
paver instead.  There should not be any ill effects since the only
reason that the zxcrypt driver would need to be bound is for the SSH
keys to be written in, which is now handled by fshost via WriteDataFile.

Tested: integration, plus manually patching on top of the linked patch.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781103
Original-Revision: 1dfecd731449ba085edddaea15d681e1fa37e009
GitOrigin-RevId: e9ce117fa19bfc3aa80a7e962b839d3b5297a2e4
Change-Id: I7c03aa54c56673ed01ca95a1a88e04f8fddd9ca9