[roll] Roll fuchsia [check-licenses] Insert proper license URL for Rust.
We want to link to the upstream copies of each license file.

The current strategy for doing this in check-licenses is somewhat rigid,
and it expects each project to only provide a single license file
or notice file. Rust provides two license files.

This CL adds a temporary bit of code to set the license URL for rust.
I'll remove this bit after designing a more flexible solution
that works for the entire repo.

Original-Bug: 118257

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782272
Original-Revision: 22e945246d9e872c167f766d7da75600e93b03dc
GitOrigin-RevId: b3516bdeefd945bdd1075917c49de6ed10f63ee0
Change-Id: I7d7d83c16eeb30275abf67e3db3699e4194cb497