[roll] Roll fuchsia [fshost] add generic config overrides
Instead of having a builder method per option we want to override, add a
generic config override method that lets tests pass in any config and
the value to set for it. This uses a trait to allow for the relevant
subset of supported config value types to be passed to the same override
function.

Original-Bug: 87511
Test: fx test fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783343
Original-Revision: 7fc509209c3b4affbd75e7e208b86ed51fc860de
GitOrigin-RevId: 3ea3fb76351834d5e190ba6aa9d7293c3e3ad0b3
Change-Id: Ieddb50104eea7e408c1bd8bda486ed0a87e12089