[roll] Roll fuchsia [build] Add support for dynamic golden file sets
This change introduces a mode of `golden_files()` in which a JSON file
of golden file comparisons may be passed instead of a literal GN list of
them. This is useful in situations in which the set of golden files
is not statically known in GN, as is the case with generating syscall
documentation pages from the FIDL syscall representations in library zx.

Since this support in `fx check-goldens` requires a building of these
manifests, this behaviour is put behind a flag

Original-Bug: 118025
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781863
Original-Revision: ee307a3e63618426e37999abb11dc57bd9b7b76a
GitOrigin-RevId: c91e6f854f4439326e050d2e5f0721f989a84eb9
Change-Id: I9348ca452fa4f1ebbaaa10ad95a51110aed82916