[roll] Roll fuchsia [check-licenses] Compress the large project.json test files.
project.json describes the workspace, and they tend to be
very very very large. One of the check-licenses tests attempts
to verify that we can accurately filter that file donw to only
the targets we care about. I committed this huge file to the
test directory, which is not ideal for several reasons
(more details in the bug).

This CL compresses the committed file, and decompresses it before
using it in the tests.

Original-Fixed: 117994
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783462
Original-Revision: 99c90c45074cc9cc63e8058899be69791f676955
GitOrigin-RevId: 2085a4b1075a98b941d2ba1ea6e7046d6b1485d7
Change-Id: I296a969c1c1f1736e3e4d7866597489689e7103b