[roll] Roll fuchsia [efi] Fix BootCpuid arithmetic error
The label `__start_BootCpuidLeaf` was defined for UEFI at an alignment
of 4 (the previous size of the contained data structure) rather than 8
(the current). As a result, we would sometime see an off-by-4 error in
the `arch::InitializeBootCpuid()` logic meant to populate this section.

This bug was consistently hit within the phys unit tests, which we
enable in this change.

Original-Fixed: 115881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783582
Original-Revision: 75e961061345f361857b6b9b6f21d8e6127b97fc
GitOrigin-RevId: 1c1fd721ba29a893531a068720286c93ce34e518
Change-Id: If43824daf58372cc7002067aa70c36c4b34920ed