[roll] Roll fuchsia [emulator][storage] remove mount and minfs dep from emulator tests
Remove the emulator test library dependency on minfs, both the host-side
tool and the filesystem itself, and the mount command line tool on
fuchsia.

Replace it with a new, small wrapper tool called run-with-logs, which
takes a binary as an argument and writes the stdout and stderr of that
binary to a provided block device in a simple format. This format is
then read by the emulator test library using another host tool,
extract-logs.

On removing the mount usage - mount in Fuchsia is currently intended to
be a debugging tool, and one which may move to /core in the future,
removing the ability for zedboot to use it. This alternative is simpler
and doesn't require an entire filesystem implementation.

Test: fx test //src/tests/disable_syscalls
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783562
Original-Revision: bcc2b7968c8546cdb3673b5a1b3aec3b97285d78
GitOrigin-RevId: 9ac07c8999a658418285e06b0cdf732e3981bd49
Change-Id: Ie56219b05c27d1fab43c894ed178896f3569f330