[roll] Roll fuchsia [check-licenses] Remove unused "diff" logic.
Previously, check-licenses had the ability to diff a given license
file against a previously released one.

It was only needed one time. If we have to go through that exercise
again, we should be able to directly diff the SPDX files instead
of needing special logic baked into check-licenses.

This CL deletes that logic, which will allow me to reorganize things
to be more flexible. If necessary, We can add back the functionality
in the future with tests for better readability and maintainability.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783843
Original-Revision: 59d7522fdfd04b7163920b5807d9871cb2116650
GitOrigin-RevId: 8ac24c5cf4f7ca095a1e159f4ff2181e7b8b4b51
Change-Id: I1f2e8e81d3160bc346d069955c70743eefe91b52