[roll] Roll fuchsia [audio][drivers][as370] Add support for concurrent dHub DMAs
We use 2 concurrent DMAs to have most of the time 2 DMAs in flight.
Without concurrent DMAs a new DMA needs to be re-enabled within
the time it takes an MTU to be drained (e.g. 333 usecs), since the
interrupt is triggered once the last MTU is transferred.
With concurrent DMAs a new DMA can be safely delayed for up
to the time it takes to drain a complete DMA (e.g. 21.3 msecs),
i.e. MTU size x dma_mtus.

Also allow the dHub thread to shutdown on a stopped thread.

Original-Bug: 107645
Test: fx test -ov syn-dhub-test
Test: Added a 19 msecs delay in the interrupt processing guarantying
the next new DMA to not be enabled in time. Without this CL, this
guarantees audio glitches for audio heard on output and recorded for
input.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782127
Original-Revision: 4ce2d78158980a4adc16b4e6afd6758f4091fab9
GitOrigin-RevId: aa950bbeaf962f486483e424be37d7ad87905801
Change-Id: I03cb094974b09bd2ae6ac7e5c5a03ad4b02b951b