[roll] Roll third_party/pigweed pw_build: Add pw-wrap-ninja helper
This commit adds a standalone Python script that wraps Ninja to provide
a more comprehensive display of the progress of the build, along with a
brief summary of the tool to the pw_build docs.

The tool works by parsing Ninja's output to build a representation of
the build status, and then outputting a richer TUI with the list of
running targets and how long they're taking. This allows one to, at a
glance, determine what targets are blocking the rest of the build or
taking an unusually long time to complete.

Original-Bug: 263976434
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/125551
Original-Revision: 21a65831d7e63b8ef73506c69627fefb6a7c6634
GitOrigin-RevId: e0c1279be86503d709c473dad8d28f06cbe51ca8
Change-Id: Idc75e1b1cfe29891f3ce063ec537c4e06bfa04fb