[roll] Roll infra/recipes [gcc] Unpack CIPD prerequisites for binutils too
The binutils configure tries to ensure that gmp and mpfr are
available when gdb is in the source tree even if it's not being
built.  So unpack the CIPD copies a second time in the binutils
source just to satisfy configure, though they won't be used.

Also add some best-practices defaults to the binutils configure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/784455
Original-Revision: c2fe2f32b5298e2c58f8f38392d7b78282fb9c19
GitOrigin-RevId: daaf9de1cfb1f08f590efa5e936987e4f6a7eb60
Change-Id: I4ae69dc5355657c589351d06442bc68ff8f28805