[roll] Roll fuchsia [starnix] binder ioctl must not return EINTR when it handled some data
The client of the binder can request both to send commands and receive
commands at the same time. When it does this, the driver waits for some
command to be available before returning. If this wait is interrupted
when the driver has handled some command from the client, it should not
return an error to let the client knows that its commands have been
handled.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778945
Original-Revision: a21d1e9ef2da59e4bcc1d68805d4bf71096dab59
GitOrigin-RevId: 536415dc3ea7dbc6f32f0d7ca3af3cad35ceb59e
Change-Id: If22256792fa84638e11a907b341d29f89714417c