[roll] Roll fuchsia [starnix] Refactor binder reference counting
The binder reference counting need to be able to notify the owner of a
reference whenever its weak or strong reference count reach 0, or
increase from 0.

The previous implementation only had reference count at the handle
level, and was relying on Rust Arc to keep the referenced object alive.
This prevented to send the correct information to the client.

This CL updates the object to have their own reference count. Moreover,
it handled sending weak count back to the owner, and handles receiving
ack from the owner for creation of weak and strong reference to an
object.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780402
Original-Revision: 125640280b096fefc500b4b4b92bee705ca37581
GitOrigin-RevId: 6843102967ef1c4e4d138f34e27b6a0755482454
Change-Id: I5892674792543f7ceb8118e4b3aadeb842960f71