[roll] Roll fuchsia [av400][dsp] Modify the way SMC accesses the dsp firmware loading address.
In order to improve the compatibility of the DSP driver, modify the
method of obtaining the loading address of the dsp firmware. In the
original program, there was an error: the value of kDspSecPowerSrt was
incorrect, and this error was not found during verification.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780943
Original-Revision: 8b4c5ec0183ac7655405ff5d8493bc4121a3d7e0
GitOrigin-RevId: a458a0e8ffcea44100290948d141be7f4a332096
Change-Id: I6053465c82b27d7a6e19944d08a34df23656db27