[roll] Roll fuchsia [identity] Update AccountHander README
This is something I've been wanting to do for a while because AH is now
our largest component and its got the most people working in it while
the documentation was woefully out of date.

Happy to make changes to the text to make it clearer and fix any errors.
Going forwards, this is about the level of detail I'd like us to
maintain for the modules so in many cases we can have a discussion about
the implementation of an epic based on the changes it wants to make to
the scope of the modules in this readme.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783508
Original-Revision: 8e54631e46888ec9915f57fe3afd3f5f9c0836ad
GitOrigin-RevId: 708a008ca700bfd69c622fed4b5b4a1e1c779387
Change-Id: I8ccbf480940e568825b39ad62f9ba0be03f11189