[roll] Roll cobalt Remove the MetadataBuilder and system profile history storage.
The 2 static methods of the MetadataBuilder that are in use are moved
to relevant locations:
- 1 is only used by the observation generator, so it is moved there
- 1 is used in several places, so it is moved to the SystemData class

The only remaining non-static use of the MetadataBuilder was in the
local aggregate storage migration code that added the system profile
to the stored protos. This rolled out in F7, and F8 is fully rolled
out now, so it should not be used very much. Instead, for late
upgrading devices, we will fall back to the current system profile
during the migration.

The files are not deleted in this change, to avoid breaking the
internal roll, but their build rules are renamed to ensure they
aren't used. A subsequent change will remove the files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/782084
Original-Revision: 34b83c73caec1e1e8129619b285ee00e1b2eb6ea
GitOrigin-RevId: c8398f9b5cb5e2f51554eb6d9139c88bf29edf86
Change-Id: I3a82ac6f114dc6b782ad28d3b978124940646e45