[roll] Roll fuchsia [emu] Add support for PBv2 with multiple VDs
This CL adds conversion code to the `ffx emu` plugin for an updated form
of PBv2 with virtual devices linked to the PBM file. This is based on
https://fxrev.dev/758802, but uses the kernel file from system_a and the
template file linked to the selected virtual device.

The virtual devices are stored in a new VirtualDeviceManifest, which is
a separate file linked to the PBM through the virtual_devices_path
field. The manifest type holds a map of names to paths to virtual
device files, and an optional recommended default device name.

Original-Bug: 116056
Original-Bug: 113144
Original-Bug: 90948

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774004
Original-Revision: 9eb14eb62d9030fe449e96171bdef35eef225f82
GitOrigin-RevId: f865e740c0563356e46ed4aa443544c1cdce4f26
Change-Id: Ic0441b9c2e2839221cdcd170810c4b42fe7ed354