[roll] Roll fuchsia [zircon] Remove bad assertion in VMAR Enumerator.
This CL removes a DEBUG_ASSERT that was in VmAddressRegionEnumerator,
that would check whether the next value in the enumerator is a mapping,
disallowing vmars.

From the bug comments:

> `itr_` points to the item after the one that was just yielded by
> `next()`, and that could either be a mapping or a vmar.

This debug assert could be triggered by issuing an operation on a vmar
that has another vmar at the end of the range. The CL adds a test that
exercises this scenario.

BUG: 118226
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785564
Original-Revision: bcbd09e4307f39680e0d6e1a2bbc26c8a9187407
GitOrigin-RevId: 3f3de32fc00bb8b57c850f23f13ee51b939597bd
Change-Id: I7c58b3c275790123b3b1f9c48b1d7dae4dfeef66