[roll] Roll fuchsia [rust-3p] Update notify to 5.0.0
This patch updates notify from 5.0.0-pre.10 to 5.0.0. It also imports
the kqueue 1.0.7, and kqueue-sys 1.0.3, which was extracted from notify
(approved in OSRB http://fxbug.dev/116738).

As part of this, this disables notify from internally using
crossbeam-channel because of
https://github.com/notify-rs/notify/issues/380, where crossbeam will
park the current thread if there's nothing on the channel. This can
cause a deadlock when running in an async executor. std::sync::mpsc
doesn't do this, so this should avoid any risk of deadlocks.

Finally, this forks the notify crate because the 5.0.0 accidentally left
out the license file. This will be fixed with
https://github.com/notify-rs/notify/issues/460.

Original-Bug: 116738
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776923
Original-Revision: f0186bf9d0fc92aea67c33a0ca4388c4bfbf4c5d
GitOrigin-RevId: 2b285a0b0e989b710bc74b65c37d58317eae3590
Change-Id: Ie819a477cf9ffa514fc1ac3e15d406b30aded2e9