[roll] Roll fuchsia [pave] Change default to --run-once
The previous default for `fx pave` and `fx pave-remote` was
`--keep-running`, so as soon as the script identified a device in
zedboot it attempted to pave and it will keep doing that cycle
forever. This creates a problem if the device is set up with USB as
the first boot because when pave issues the reboot command it
reboots back into zedboot and starts paving all over again. The
problem could be prevented by passing the -1 argument to the script
so that it only attempts to pave once, but many users (especially
new ones) don't know about this argument and don't expect the
continuous paving behavior.

This CL changes the default to `--run-once`. It also changes the flag
from `-1` to `--run-once` in order to be a bit more dscriptive.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780646
Original-Revision: 4d2dbe250dcad27169e210436792b02f9c9ee8f7
GitOrigin-RevId: 0642a67e1db15ab47bf4ff688adc30f937104d7b
Change-Id: I5e9673d47ac24fb4ea90a0a1e8e40cc2519ee34f