[roll] Roll fuchsia [fidl][screenshot] Add new TakeFile method to the screenshot FIDL API
When used, the screenshot will be sent as a |fuchsia.io.File| over
the using the file channel in the |Screenshot| response.
This flow typically useful if the client is on
the host and does not support VMOs, as is the case for ffx tools.
Both in gfx and in flatland a map is added to hold the vmofile and
the channel observer.
Once the channel is closed, the entry is removed from the map.

Test: UT
Original-Bug: 109048

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762865
Original-Revision: aa7aed4ec95fbb406da097fd7ef5b46c4cce30cc
GitOrigin-RevId: dbf83f0eb53b946191a9cfaa7e98062fe24d6eed
Change-Id: I1f8489673250464432528d6bec38ad124f5ea86a