[roll] Roll fuchsia [moniker] Remove `rep` field from ChildMoniker
ChildMoniker stores the display representation of the moniker
as a string, in addition to the separate strings for the
collection and name.

This uses memory unnecessarily. Construct the display
representation when it is needed using the `ToString` or
`Display` trait.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785602
Original-Revision: 5ec2ecfe7c5f8a92c2ae71d2d255241a39a1380b
GitOrigin-RevId: 2c83364237fc0a1af47a05bc4f2955db2ebf7fed
Change-Id: If501dcf43cf438f71896416239f59dfb7a46795d